ASQ dataset

This dataset is associated with the following objects:

Details

response_asq a data.frame containing raw response data of 751 participants and 41 variables.

prosettaid. participant IDs.

EDANX01 -- MASQ11. response to items.

itemmap_asq a data.frame containing the item map, describing the items in each instrument.

item_order item numeric IDs. This column refers to the column item_order in anchor item attributes.

instrument the instrument ID that each item belongs to.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

item_name new item ID strings to be used in the combined scale.

ncat the number of response categories.

min_score the minimum score of the item.

reverse whether the item data has been reverse-scored. 1 indicates the item has been reverse-scored, and 0 indicates the item has not been reverse-scored.

scores a comma-separated string representing all possible score values from the item.

anchor_asq a data.frame containing anchor item parameters for 29 items.

item_order item numeric IDs.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

a the discrimination (slope) parameter for the graded response model.

cb1 - cb4 the boundaries between each category-pair for the graded response model.

ncat the number of response categories.

data_asq a PROsetta_data object containing the datasets above. See loadData for creating PROsetta_data objects.

Examples

## load datasets into a PROsetta_data object
data_asq <- loadData(
  response = response_asq,
  itemmap  = itemmap_asq,
  anchor   = anchor_asq
)

#> 'reverse' column is present in item map, and no items are marked as reversed

#> item_id guessed as   : item_id
#> person_id guessed as : prosettaid
#> scale_id guessed as  : instrument

## run descriptive statistics
runDescriptive(data_asq)

#>           n mean   sd median trimmed  mad min max range skew kurtosis   se
#> EDANX01 751 1.50 0.84      1    1.32 0.00   1   5     4 1.76     2.72 0.03
#> EDANX02 751 1.43 0.75      1    1.25 0.00   1   5     4 1.89     3.39 0.03
#> EDANX03 750 1.43 0.81      1    1.24 0.00   1   5     4 2.01     3.68 0.03
#> EDANX05 751 1.86 1.03      2    1.70 1.48   1   5     4 1.04     0.25 0.04
#> EDANX07 751 1.46 0.89      1    1.26 0.00   1   5     4 1.95     3.10 0.03
#> EDANX08 751 1.57 0.95      1    1.37 0.00   1   5     4 1.67     2.08 0.03
#> EDANX12 750 1.99 0.98      2    1.90 1.48   1   5     4 0.62    -0.49 0.04
#> EDANX13 751 1.52 0.83      1    1.35 0.00   1   5     4 1.56     1.78 0.03
#> EDANX16 750 1.67 0.95      1    1.50 0.00   1   5     4 1.34     1.07 0.03
#> EDANX18 751 1.41 0.79      1    1.22 0.00   1   5     4 2.06     3.97 0.03
#> EDANX20 751 1.65 0.91      1    1.48 0.00   1   5     4 1.48     1.87 0.03
#> EDANX21 750 1.87 0.96      2    1.75 1.48   1   5     4 0.91     0.19 0.04
#> EDANX24 751 1.71 1.06      1    1.50 0.00   1   5     4 1.30     0.54 0.04
#> EDANX26 751 1.79 0.98      1    1.64 0.00   1   5     4 1.01     0.08 0.04
#> EDANX27 751 1.51 0.87      1    1.32 0.00   1   5     4 1.77     2.60 0.03
#> EDANX30 751 2.04 1.05      2    1.90 1.48   1   5     4 0.71    -0.41 0.04
#> EDANX33 750 1.24 0.61      1    1.08 0.00   1   5     4 3.01    10.20 0.02
#> EDANX37 750 1.91 1.07      2    1.75 1.48   1   5     4 1.02     0.28 0.04
#> EDANX40 750 1.40 0.75      1    1.21 0.00   1   5     4 2.03     3.97 0.03
#> EDANX41 751 1.55 0.92      1    1.36 0.00   1   5     4 1.68     2.14 0.03
#> EDANX44 751 1.53 0.86      1    1.36 0.00   1   5     4 1.60     1.88 0.03
#> EDANX46 751 1.80 0.94      2    1.67 1.48   1   5     4 0.94     0.01 0.03
#> EDANX47 751 1.90 1.03      2    1.75 1.48   1   5     4 0.92     0.01 0.04
#> EDANX48 750 1.83 1.01      2    1.68 1.48   1   5     4 1.10     0.52 0.04
#> EDANX49 751 2.41 1.21      2    2.31 1.48   1   5     4 0.40    -0.84 0.04
#> EDANX51 751 2.03 1.06      2    1.90 1.48   1   5     4 0.68    -0.51 0.04
#> EDANX53 751 1.83 0.98      2    1.69 1.48   1   5     4 0.98     0.13 0.04
#> EDANX54 751 2.06 1.07      2    1.93 1.48   1   5     4 0.65    -0.55 0.04
#> EDANX55 751 1.55 0.86      1    1.38 0.00   1   5     4 1.55     1.82 0.03
#> MASQ1   751 1.34 0.65      1    1.20 0.00   1   5     4 2.16     4.87 0.02
#> MASQ2   751 1.40 0.75      1    1.22 0.00   1   5     4 2.05     3.95 0.03
#> MASQ3   751 1.63 0.83      1    1.48 0.00   1   5     4 1.40     1.77 0.03
#> MASQ4   751 1.69 0.88      1    1.54 0.00   1   5     4 1.37     1.67 0.03
#> MASQ5   751 1.20 0.56      1    1.06 0.00   1   5     4 3.75    16.88 0.02
#> MASQ6   751 1.57 0.82      1    1.41 0.00   1   5     4 1.50     1.91 0.03
#> MASQ7   751 1.73 0.94      1    1.57 0.00   1   5     4 1.31     1.27 0.03
#> MASQ8   751 1.70 0.96      1    1.52 0.00   1   5     4 1.41     1.41 0.03
#> MASQ9   751 1.35 0.69      1    1.19 0.00   1   5     4 2.28     5.58 0.03
#> MASQ10  750 1.69 0.97      1    1.50 0.00   1   5     4 1.46     1.49 0.04
#> MASQ11  750 2.07 1.02      2    1.94 1.48   1   5     4 0.79     0.04 0.04

## run item parameter calibration on the response data, linking to the anchor item parameters
runLinking(data_asq, method = "FIXEDPAR")

#> response data has 40 items

#> performing 1D fixed parameter calibration, using anchor data

#> anchor has 29 items * 5 parameters = 145 parameters

#> anchor parameters applied to par_layout

#> Iteration: 1, Log-Lik: -24053.720, Max-Change: 2.73844Iteration: 2, Log-Lik: -23453.207, Max-Change: 0.72611Iteration: 3, Log-Lik: -23439.758, Max-Change: 0.41693Iteration: 4, Log-Lik: -23438.315, Max-Change: 0.66336Iteration: 5, Log-Lik: -23437.791, Max-Change: 0.50564Iteration: 6, Log-Lik: -23437.683, Max-Change: 0.13194Iteration: 7, Log-Lik: -23437.680, Max-Change: 0.08704Iteration: 8, Log-Lik: -23437.677, Max-Change: 0.00500Iteration: 9, Log-Lik: -23437.677, Max-Change: 0.00021Iteration: 10, Log-Lik: -23437.677, Max-Change: 0.00014Iteration: 11, Log-Lik: -23437.677, Max-Change: 0.00014Iteration: 12, Log-Lik: -23437.677, Max-Change: 0.00005

#> $constants
#> [1] NA
#> 
#> $ipar_linked
#>                 a          b1         b2       b3       b4
#> EDANX01 3.6021544  0.34159048 1.08952607 1.960101 2.698704
#> EDANX02 3.4504105  0.49115372 1.33003913 2.155174 2.880245
#> EDANX03 3.3365573  0.62131513 1.24324827 2.061498 2.758621
#> EDANX05 3.3552816 -0.18966149 0.59808268 1.574890 2.445823
#> EDANX07 3.5509265  0.53941950 1.04602617 1.865002 2.384703
#> EDANX08 2.8482390  0.36804661 1.06412460 1.853262 2.544713
#> EDANX12 2.7344963 -0.56874342 0.47212997 1.671720 2.969534
#> EDANX13 1.4299890  0.32031811 1.38525972 2.740602 4.657334
#> EDANX16 1.8026250 -0.03826752 0.87866300 2.112580 3.045698
#> EDANX18 2.9857713  0.56615423 1.30206069 2.149592 3.060772
#> EDANX20 1.7235167  0.01038630 1.16848655 2.202962 3.108540
#> EDANX21 2.2320871 -0.30039122 0.73730822 1.859789 2.924760
#> EDANX24 1.8300229  0.33186966 0.98209963 1.887270 3.023501
#> EDANX26 1.9506502 -0.02653687 0.87921023 2.014506 3.240672
#> EDANX27 2.8602293  0.43614207 1.12685718 2.029540 2.777844
#> EDANX30 3.0318433 -0.51540331 0.31555222 1.351650 2.299683
#> EDANX33 2.5819082  1.14704504 1.81665694 2.716464 3.589089
#> EDANX37 1.7091745 -0.03958087 0.88362788 2.001631 3.104371
#> EDANX40 3.8831993  0.48588040 1.26317743 2.111117 2.898513
#> EDANX41 3.6595132  0.36439509 1.03380800 1.780490 2.621156
#> EDANX44 1.2704961  0.57192503 1.45897673 2.677899 3.972340
#> EDANX46 3.3981373 -0.21663501 0.63213040 1.644448 2.731239
#> EDANX47 2.5365096 -0.23896204 0.67780928 1.795603 2.900682
#> EDANX48 3.0433412 -0.33156350 0.55586628 1.455912 2.339870
#> EDANX49 1.5157005 -0.82984231 0.08690911 1.167730 2.379728
#> EDANX51 2.4116611 -0.46182440 0.39835327 1.363589 2.402821
#> EDANX53 3.6561060 -0.23177839 0.59522713 1.563458 2.499054
#> EDANX54 3.3503266 -0.50940366 0.31066694 1.250223 2.296618
#> EDANX55 3.1273340  0.07436260 0.94088815 1.847720 2.775704
#> MASQ1   2.4553590  0.62790093 1.76305616 2.556934 4.346750
#> MASQ2   0.7737667  1.22523358 3.13716634 4.695677 7.982062
#> MASQ3   2.9565421 -0.04681016 1.19269265 1.877235 3.041975
#> MASQ4   3.0397128 -0.12152427 1.05359329 1.773160 2.642031
#> MASQ5   1.2272524  1.63114582 3.27898650 3.861736 4.859649
#> MASQ6   1.0794230  0.24091646 2.01113983 3.214757 5.576669
#> MASQ7   2.2877395 -0.12364435 1.01625412 1.752663 2.756741
#> MASQ8   2.1625206 -0.01949997 1.07289020 1.777269 2.794919
#> MASQ9   1.1514947  0.98152540 2.52516400 3.723844 5.312664
#> MASQ10  2.3261588  0.02420731 1.06489797 1.643662 2.644612
#> MASQ11  0.8225668 -1.13444184 1.01296265 2.781961 4.912505
#> 
#> $ipar_anchor
#>                a          b1         b2       b3       b4
#> EDANX01 3.602154  0.34159048 1.08952607 1.960101 2.698704
#> EDANX02 3.450410  0.49115372 1.33003913 2.155174 2.880245
#> EDANX03 3.336557  0.62131513 1.24324827 2.061498 2.758621
#> EDANX05 3.355282 -0.18966149 0.59808268 1.574890 2.445823
#> EDANX07 3.550927  0.53941950 1.04602617 1.865002 2.384703
#> EDANX08 2.848239  0.36804661 1.06412460 1.853262 2.544713
#> EDANX12 2.734496 -0.56874342 0.47212997 1.671720 2.969534
#> EDANX13 1.429989  0.32031811 1.38525972 2.740602 4.657334
#> EDANX16 1.802625 -0.03826752 0.87866300 2.112580 3.045698
#> EDANX18 2.985771  0.56615423 1.30206069 2.149592 3.060772
#> EDANX20 1.723517  0.01038630 1.16848655 2.202962 3.108540
#> EDANX21 2.232087 -0.30039122 0.73730822 1.859789 2.924760
#> EDANX24 1.830023  0.33186966 0.98209963 1.887270 3.023501
#> EDANX26 1.950650 -0.02653687 0.87921023 2.014506 3.240672
#> EDANX27 2.860229  0.43614207 1.12685718 2.029540 2.777844
#> EDANX30 3.031843 -0.51540331 0.31555223 1.351650 2.299683
#> EDANX33 2.581908  1.14704504 1.81665694 2.716464 3.589089
#> EDANX37 1.709175 -0.03958087 0.88362788 2.001631 3.104371
#> EDANX40 3.883199  0.48588040 1.26317743 2.111117 2.898513
#> EDANX41 3.659513  0.36439509 1.03380800 1.780490 2.621156
#> EDANX44 1.270496  0.57192503 1.45897673 2.677899 3.972340
#> EDANX46 3.398137 -0.21663501 0.63213040 1.644448 2.731239
#> EDANX47 2.536510 -0.23896204 0.67780928 1.795603 2.900682
#> EDANX48 3.043341 -0.33156350 0.55586628 1.455912 2.339870
#> EDANX49 1.515701 -0.82984231 0.08690911 1.167730 2.379728
#> EDANX51 2.411661 -0.46182440 0.39835327 1.363589 2.402821
#> EDANX53 3.656106 -0.23177839 0.59522713 1.563458 2.499054
#> EDANX54 3.350327 -0.50940366 0.31066694 1.250223 2.296618
#> EDANX55 3.127334  0.07436260 0.94088815 1.847720 2.775704
#> 
#> $method
#> [1] "FIXEDPAR"
#>