DEP dataset

This dataset is associated with the following objects:

Details

response_dep a data.frame containing raw response data of 747 participants and 49 variables.

prosettaid. participant IDs.

EDDEP04 -- CESD20. response to items.

itemmap_dep a data.frame containing the item map, describing the items in each instrument.

item_order item numeric IDs. This column refers to the column item_order in anchor item parameters.

instrument the instrument ID that each item belongs to.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

item_name new item ID strings to be used in the combined scale.

ncat the number of response categories.

min_score the minimum score of the item.

reverse whether the item data has been reverse-scored. 1 indicates the item has been reverse-scored, and 0 indicates the item has not been reverse-scored.

scores a comma-separated string representing all possible score values from the item.

anchor_dep a data.frame containing anchor item parameters for 28 items.

item_order item numeric IDs.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

a the discrimination (slope) parameter for the graded response model.

cb1 - cb4 the boundaries between each category-pair for the graded response model.

ncat the number of response categories.

data_dep a PROsetta_data object containing the datasets above. See loadData for creating PROsetta_data objects.

Examples

## load datasets into a PROsetta_data object
data_dep <- loadData(
  response = response_dep,
  itemmap  = itemmap_dep,
  anchor   = anchor_dep
)

#> some items are marked as reversed items in 'reverse' column in item map

#> assuming the response data is already reverse coded

#> item_id guessed as   : item_id
#> person_id guessed as : prosettaid
#> scale_id guessed as  : instrument

## run descriptive statistics
runDescriptive(data_dep)

#>           n mean   sd median trimmed  mad min max range skew kurtosis   se
#> EDDEP04 747 1.52 0.94      1    1.30 0.00   1   5     4 1.91     3.01 0.03
#> EDDEP05 746 1.62 0.99      1    1.42 0.00   1   5     4 1.54     1.50 0.04
#> EDDEP06 746 1.56 0.94      1    1.37 0.00   1   5     4 1.66     2.01 0.03
#> EDDEP07 747 1.78 1.05      1    1.59 0.00   1   5     4 1.23     0.59 0.04
#> EDDEP09 747 1.56 0.91      1    1.38 0.00   1   5     4 1.62     1.98 0.03
#> EDDEP14 747 1.69 1.00      1    1.51 0.00   1   5     4 1.38     1.12 0.04
#> EDDEP17 747 2.08 1.04      2    1.95 1.48   1   5     4 0.66    -0.40 0.04
#> EDDEP19 747 1.47 0.90      1    1.26 0.00   1   5     4 2.02     3.56 0.03
#> EDDEP21 746 1.77 0.99      1    1.61 0.00   1   5     4 1.18     0.70 0.04
#> EDDEP22 746 1.68 1.02      1    1.50 0.00   1   5     4 1.42     1.26 0.04
#> EDDEP23 747 1.91 1.09      2    1.73 1.48   1   5     4 0.98     0.01 0.04
#> EDDEP26 746 2.05 1.10      2    1.90 1.48   1   5     4 0.81    -0.18 0.04
#> EDDEP27 744 1.66 0.99      1    1.47 0.00   1   5     4 1.43     1.23 0.04
#> EDDEP28 747 1.89 1.09      1    1.70 0.00   1   5     4 1.05     0.16 0.04
#> EDDEP29 747 1.86 1.06      1    1.69 0.00   1   5     4 1.06     0.20 0.04
#> EDDEP30 747 1.83 0.98      2    1.69 1.48   1   5     4 1.02     0.34 0.04
#> EDDEP31 746 2.04 1.12      2    1.89 1.48   1   5     4 0.76    -0.35 0.04
#> EDDEP35 747 1.86 1.04      1    1.70 0.00   1   5     4 1.01     0.13 0.04
#> EDDEP36 747 2.05 1.07      2    1.90 1.48   1   5     4 0.79    -0.14 0.04
#> EDDEP39 747 1.29 0.74      1    1.09 0.00   1   5     4 2.88     8.35 0.03
#> EDDEP41 747 1.54 0.93      1    1.33 0.00   1   5     4 1.77     2.40 0.03
#> EDDEP42 747 1.80 1.00      1    1.65 0.00   1   5     4 1.09     0.39 0.04
#> EDDEP44 747 1.61 0.90      1    1.44 0.00   1   5     4 1.36     0.98 0.03
#> EDDEP45 747 1.69 0.91      1    1.55 0.00   1   5     4 1.19     0.76 0.03
#> EDDEP46 747 2.01 1.09      2    1.85 1.48   1   5     4 0.84    -0.16 0.04
#> EDDEP48 747 1.67 1.04      1    1.47 0.00   1   5     4 1.44     1.18 0.04
#> EDDEP50 746 1.72 0.99      1    1.55 0.00   1   5     4 1.30     1.01 0.04
#> EDDEP54 747 1.95 1.11      2    1.78 1.48   1   5     4 0.95    -0.06 0.04
#> CESD1   746 1.31 0.62      1    1.16 0.00   1   4     3 2.11     4.03 0.02
#> CESD2   747 1.26 0.61      1    1.11 0.00   1   4     3 2.61     6.83 0.02
#> CESD3   747 1.35 0.77      1    1.14 0.00   1   4     3 2.23     3.95 0.03
#> CESD4   747 1.72 1.09      1    1.53 0.00   1   4     3 1.17    -0.20 0.04
#> CESD5   747 1.53 0.81      1    1.38 0.00   1   4     3 1.41     1.08 0.03
#> CESD6   747 1.49 0.87      1    1.29 0.00   1   4     3 1.72     1.90 0.03
#> CESD7   747 1.61 0.89      1    1.43 0.00   1   4     3 1.38     0.93 0.03
#> CESD8   747 1.98 1.13      2    1.85 1.48   1   4     3 0.70    -1.00 0.04
#> CESD9   747 1.39 0.81      1    1.18 0.00   1   4     3 2.11     3.42 0.03
#> CESD10  747 1.23 0.57      1    1.09 0.00   1   4     3 2.58     6.27 0.02
#> CESD11  747 1.99 1.03      2    1.86 1.48   1   4     3 0.68    -0.78 0.04
#> CESD12  746 1.73 0.99      1    1.56 0.00   1   4     3 1.08    -0.13 0.04
#> CESD13  745 1.59 0.84      1    1.43 0.00   1   4     3 1.32     0.88 0.03
#> CESD14  745 1.53 0.90      1    1.33 0.00   1   4     3 1.62     1.48 0.03
#> CESD15  744 1.33 0.64      1    1.18 0.00   1   4     3 2.08     4.07 0.02
#> CESD16  745 1.71 1.00      1    1.54 0.00   1   4     3 1.09    -0.19 0.04
#> CESD17  745 1.18 0.53      1    1.02 0.00   1   4     3 3.26    10.59 0.02
#> CESD18  743 1.55 0.84      1    1.38 0.00   1   4     3 1.46     1.30 0.03
#> CESD19  742 1.36 0.71      1    1.20 0.00   1   4     3 2.10     4.02 0.03
#> CESD20  743 1.76 0.90      2    1.62 1.48   1   4     3 1.02     0.14 0.03

## run item parameter calibration on the response data, linking to the anchor item parameters
runLinking(data_dep, method = "FIXEDPAR")

#> response data has 48 items

#> performing 1D fixed parameter calibration, using anchor data

#> anchor has 28 items * 5 parameters = 140 parameters

#> anchor parameters applied to par_layout

#> Iteration: 1, Log-Lik: -27522.223, Max-Change: 2.50451Iteration: 2, Log-Lik: -26639.786, Max-Change: 0.49343Iteration: 3, Log-Lik: -26614.123, Max-Change: 0.21027Iteration: 4, Log-Lik: -26610.945, Max-Change: 0.09447Iteration: 5, Log-Lik: -26610.482, Max-Change: 0.06020Iteration: 6, Log-Lik: -26610.409, Max-Change: 0.01176Iteration: 7, Log-Lik: -26610.405, Max-Change: 0.00876Iteration: 8, Log-Lik: -26610.401, Max-Change: 0.00149Iteration: 9, Log-Lik: -26610.400, Max-Change: 0.00050Iteration: 10, Log-Lik: -26610.400, Max-Change: 0.00029Iteration: 11, Log-Lik: -26610.400, Max-Change: 0.00011Iteration: 12, Log-Lik: -26610.400, Max-Change: 0.00003

#> $constants
#> [1] NA
#> 
#> $ipar_linked
#>                a          b1        b2       b3       b4
#> EDDEP04 4.261422  0.40106943 0.9756732 1.696300 2.444072
#> EDDEP05 3.931743  0.30494182 0.9130961 1.593476 2.411682
#> EDDEP06 4.144759  0.35011299 0.9153482 1.678203 2.470526
#> EDDEP07 2.801804  0.14774854 0.7723478 1.602715 2.538057
#> EDDEP09 3.657433  0.31195821 0.9818088 1.782108 2.571127
#> EDDEP14 2.333381  0.18599340 0.9473173 1.728770 2.632643
#> EDDEP17 3.274033 -0.49845044 0.4059439 1.413052 2.375459
#> EDDEP19 3.240973  0.46049359 1.0344267 1.833595 2.514716
#> EDDEP21 2.736104  0.07245992 0.8097820 1.803067 2.673441
#> EDDEP22 3.970028  0.20379963 0.7954889 1.648707 2.295496
#> EDDEP23 2.564431 -0.03841113 0.6926952 1.652820 2.583629
#> EDDEP26 3.093368 -0.35762208 0.4124999 1.403863 2.223961
#> EDDEP27 2.920056  0.20434012 0.8909159 1.654652 2.528368
#> EDDEP28 2.588339 -0.07908821 0.6326205 1.477330 2.327715
#> EDDEP29 4.342918 -0.11730304 0.5977487 1.428166 2.272495
#> EDDEP30 2.612846 -0.02337743 0.8683854 1.864303 2.826340
#> EDDEP31 3.182866 -0.26089392 0.3967655 1.305464 2.133989
#> EDDEP35 3.105857  0.04371372 0.7223523 1.638758 2.471487
#> EDDEP36 3.483012 -0.53588456 0.3475707 1.346781 2.354790
#> EDDEP39 3.131213  0.91802061 1.4813240 2.163996 2.856377
#> EDDEP41 4.454157  0.55838282 1.0742430 1.779346 2.530080
#> EDDEP42 2.364413  0.21006534 0.9870936 1.905901 2.933758
#> EDDEP44 2.549164  0.19350024 1.0117460 2.013109 3.126494
#> EDDEP45 2.833605  0.14071213 0.9065001 1.846097 2.875194
#> EDDEP46 2.380628 -0.45788346 0.4779996 1.545663 2.631512
#> EDDEP48 3.185244  0.19814448 0.7819065 1.525815 2.324082
#> EDDEP50 2.018099 -0.05044210 0.9258678 1.999516 2.965507
#> EDDEP54 2.685300 -0.29880847 0.4234598 1.357851 2.307647
#> CESD1   2.074442  0.87588000 1.9208517 3.063586       NA
#> CESD2   1.262450  1.38731970 2.6695226 3.720540       NA
#> CESD3   3.512477  0.83271884 1.3159735 1.948794       NA
#> CESD4   1.118259  0.64882022 1.3786370 2.081167       NA
#> CESD5   1.604924  0.42935043 1.5261799 2.723723       NA
#> CESD6   3.634668  0.49267515 1.1756152 1.729146       NA
#> CESD7   1.827676  0.28707337 1.3677916 2.134161       NA
#> CESD8   1.341784 -0.06693047 0.8227682 1.619581       NA
#> CESD9   3.002587  0.74752445 1.3741501 1.855369       NA
#> CESD10  2.060394  1.17178033 2.0425563 3.268036       NA
#> CESD11  1.076593 -0.46319695 0.9471572 2.159803       NA
#> CESD12  2.229316  0.16859288 0.9449513 1.736562       NA
#> CESD13  1.288445  0.34213331 1.6959381 2.915162       NA
#> CESD14  2.176407  0.49140714 1.2914091 1.864275       NA
#> CESD15  1.396543  0.96457795 2.3205020 3.603940       NA
#> CESD16  2.132818  0.27238860 0.9216832 1.808276       NA
#> CESD17  1.718811  1.60657580 2.3169977 3.469721       NA
#> CESD18  2.812159  0.26144281 1.2484443 1.984112       NA
#> CESD19  1.833676  0.78350053 1.8751255 2.638592       NA
#> CESD20  1.490735 -0.14037590 1.2558582 2.296856       NA
#> 
#> $ipar_anchor
#>                a          b1        b2       b3       b4
#> EDDEP04 4.261422  0.40106943 0.9756732 1.696300 2.444072
#> EDDEP05 3.931743  0.30494182 0.9130961 1.593476 2.411682
#> EDDEP06 4.144759  0.35011299 0.9153482 1.678203 2.470526
#> EDDEP07 2.801804  0.14774854 0.7723478 1.602715 2.538057
#> EDDEP09 3.657433  0.31195821 0.9818088 1.782108 2.571127
#> EDDEP14 2.333381  0.18599340 0.9473173 1.728770 2.632643
#> EDDEP17 3.274033 -0.49845044 0.4059439 1.413052 2.375459
#> EDDEP19 3.240973  0.46049359 1.0344267 1.833595 2.514716
#> EDDEP21 2.736104  0.07245992 0.8097820 1.803067 2.673441
#> EDDEP22 3.970028  0.20379963 0.7954889 1.648707 2.295496
#> EDDEP23 2.564431 -0.03841113 0.6926952 1.652820 2.583629
#> EDDEP26 3.093368 -0.35762208 0.4124999 1.403863 2.223961
#> EDDEP27 2.920056  0.20434012 0.8909159 1.654652 2.528368
#> EDDEP28 2.588339 -0.07908821 0.6326205 1.477330 2.327715
#> EDDEP29 4.342918 -0.11730304 0.5977487 1.428166 2.272495
#> EDDEP30 2.612846 -0.02337743 0.8683854 1.864303 2.826340
#> EDDEP31 3.182866 -0.26089392 0.3967655 1.305464 2.133989
#> EDDEP35 3.105857  0.04371372 0.7223523 1.638758 2.471487
#> EDDEP36 3.483012 -0.53588456 0.3475707 1.346781 2.354790
#> EDDEP39 3.131213  0.91802061 1.4813240 2.163996 2.856377
#> EDDEP41 4.454157  0.55838282 1.0742430 1.779346 2.530080
#> EDDEP42 2.364413  0.21006534 0.9870936 1.905901 2.933758
#> EDDEP44 2.549164  0.19350024 1.0117460 2.013109 3.126494
#> EDDEP45 2.833605  0.14071213 0.9065001 1.846097 2.875194
#> EDDEP46 2.380628 -0.45788346 0.4779996 1.545663 2.631512
#> EDDEP48 3.185244  0.19814448 0.7819065 1.525815 2.324082
#> EDDEP50 2.018099 -0.05044210 0.9258678 1.999516 2.965507
#> EDDEP54 2.685300 -0.29880847 0.4234598 1.357851 2.307647
#> 
#> $method
#> [1] "FIXEDPAR"
#>