Plot frequency distribution

This is an extension of plot to visualize frequency distribution from PROsetta_data object.

# S4 method for PROsetta_data,ANY
plot(
  x,
  y,
  scale_id = "combined",
  filename = NULL,
  title = NULL,
  xlim = NULL,
  color = "blue",
  nbar = 20,
  rug = FALSE,
  filetype = "pdf",
  savefile = FALSE,
  bg = "white",
  width = 6,
  height = 6,
  pointsize = 12
)

Arguments

x

a PROsetta_data object.

y

unused argument, exists for compatibility with plot in the base R package.

scale_id

scale ID to plot. combined represents the combined scale. (default = combined)

filename

the filename to write if savefile argument is TRUE.

title

the title of the figure.

xlim

the range of scores to plot.

color

the color to fill the histogram. (default = blue)

nbar

the number of histogram bars. (default = 20)

rug

if TRUE, display the actual distribution of scores below each bar. (default = FALSE)

filetype

the type of file to write if savefile argument is TRUE. Accepts pdf, jpeg, png, and tiff. (default = pdf)

savefile

if TRUE, save the figure as a file. (default = FALSE)

bg

the background color to use when saving the figure as a file. (default = white)

width

the width of the plot. (default = 6)

height

the height of the plot. (default = 6)

pointsize

the point size to use when saving the figure as a file. (default = 12)

Examples

plot(data_asq)
plot(data_asq, scale_id = 1)
plot(data_asq, scale_id = 2)