Run a confirmatory factor analysis

runCFA is a function to perform a one-factor confirmatory factor analysis (CFA) to test unidimensionality.

runCFA(data, estimator = "WLSMV", std.lv = TRUE, scalewise = FALSE, ...)

Arguments

data
a PROsetta_data object. See loadData for loading a dataset.

estimator
the estimator to be used. Passed onto cfa in 'lavaan' package. (default = WLSMV)

std.lv
if TRUE, the metric of the latent variable is determined by fixing their (residual) variances to 1.0. If FALSE, the metric of each latent variable is determined by fixing the factor loading of the first indicator to 1.0. Passed onto cfa. (default = TRUE)

scalewise
if TRUE, run analysis for each scale as well as for the combined scale. If FALSE, run analysis only for the combined scale. (default = FALSE)

...
additional arguments to pass onto cfa.

Value

runCFA returns a list containing the CFA results.

Examples

# \donttest{
out_cfa <- runCFA(data_asq, scalewise = TRUE)

#> Warning: lavaan WARNING:
#>     The variance-covariance matrix of the estimated parameters (vcov)
#>     does not appear to be positive definite! The smallest eigenvalue
#>     (= 1.920901e-17) is close to zero. This may be a symptom that the
#>     model is not identified.

#> Warning: lavaan WARNING:
#>     The variance-covariance matrix of the estimated parameters (vcov)
#>     does not appear to be positive definite! The smallest eigenvalue
#>     (= 8.855631e-18) is close to zero. This may be a symptom that the
#>     model is not identified.

#> Warning: lavaan WARNING:
#>     The variance-covariance matrix of the estimated parameters (vcov)
#>     does not appear to be positive definite! The smallest eigenvalue
#>     (= -3.378904e-17) is smaller than zero. This may be a symptom that
#>     the model is not identified.

lavaan::summary(out_cfa$`1`       , fit.measures = TRUE, standardized = TRUE, estimates = FALSE)

#> lavaan 0.6-7 ended normally after 20 iterations
#> 
#>   Estimator                                       DWLS
#>   Optimization method                           NLMINB
#>   Number of free parameters                        145
#>                                                       
#>                                                   Used       Total
#>   Number of observations                           744         751
#>                                                                   
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                               736.918    1209.234
#>   Degrees of freedom                               377         377
#>   P-value (Chi-square)                           0.000       0.000
#>   Scaling correction factor                                  0.703
#>   Shift parameter                                          161.448
#>        simple second-order correction                             
#> 
#> Model Test Baseline Model:
#> 
#>   Test statistic                            295948.962   48057.755
#>   Degrees of freedom                               406         406
#>   P-value                                        0.000       0.000
#>   Scaling correction factor                                  6.202
#> 
#> User Model versus Baseline Model:
#> 
#>   Comparative Fit Index (CFI)                    0.999       0.983
#>   Tucker-Lewis Index (TLI)                       0.999       0.981
#>                                                                   
#>   Robust Comparative Fit Index (CFI)                            NA
#>   Robust Tucker-Lewis Index (TLI)                               NA
#> 
#> Root Mean Square Error of Approximation:
#> 
#>   RMSEA                                          0.036       0.055
#>   90 Percent confidence interval - lower         0.032       0.051
#>   90 Percent confidence interval - upper         0.040       0.058
#>   P-value RMSEA <= 0.05                          1.000       0.016
#>                                                                   
#>   Robust RMSEA                                                  NA
#>   90 Percent confidence interval - lower                        NA
#>   90 Percent confidence interval - upper                        NA
#> 
#> Standardized Root Mean Square Residual:
#> 
#>   SRMR                                           0.035       0.035

lavaan::summary(out_cfa$`2`       , fit.measures = TRUE, standardized = TRUE, estimates = FALSE)

#> lavaan 0.6-7 ended normally after 18 iterations
#> 
#>   Estimator                                       DWLS
#>   Optimization method                           NLMINB
#>   Number of free parameters                         55
#>                                                       
#>                                                   Used       Total
#>   Number of observations                           750         751
#>                                                                   
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                               716.876     916.917
#>   Degrees of freedom                                44          44
#>   P-value (Chi-square)                           0.000       0.000
#>   Scaling correction factor                                  0.794
#>   Shift parameter                                           14.139
#>        simple second-order correction                             
#> 
#> Model Test Baseline Model:
#> 
#>   Test statistic                             29341.781   14025.020
#>   Degrees of freedom                                55          55
#>   P-value                                        0.000       0.000
#>   Scaling correction factor                                  2.096
#> 
#> User Model versus Baseline Model:
#> 
#>   Comparative Fit Index (CFI)                    0.977       0.938
#>   Tucker-Lewis Index (TLI)                       0.971       0.922
#>                                                                   
#>   Robust Comparative Fit Index (CFI)                            NA
#>   Robust Tucker-Lewis Index (TLI)                               NA
#> 
#> Root Mean Square Error of Approximation:
#> 
#>   RMSEA                                          0.143       0.163
#>   90 Percent confidence interval - lower         0.134       0.154
#>   90 Percent confidence interval - upper         0.152       0.172
#>   P-value RMSEA <= 0.05                          0.000       0.000
#>                                                                   
#>   Robust RMSEA                                                  NA
#>   90 Percent confidence interval - lower                        NA
#>   90 Percent confidence interval - upper                        NA
#> 
#> Standardized Root Mean Square Residual:
#> 
#>   SRMR                                           0.110       0.110

lavaan::summary(out_cfa$`combined`, fit.measures = TRUE, standardized = TRUE, estimates = FALSE)

#> lavaan 0.6-7 ended normally after 22 iterations
#> 
#>   Estimator                                       DWLS
#>   Optimization method                           NLMINB
#>   Number of free parameters                        200
#>                                                       
#>                                                   Used       Total
#>   Number of observations                           743         751
#>                                                                   
#> Model Test User Model:
#>                                               Standard      Robust
#>   Test Statistic                              3095.729    3755.056
#>   Degrees of freedom                               740         740
#>   P-value (Chi-square)                           0.000       0.000
#>   Scaling correction factor                                  0.921
#>   Shift parameter                                          392.498
#>        simple second-order correction                             
#> 
#> Model Test Baseline Model:
#> 
#>   Test statistic                            444178.742   62407.029
#>   Degrees of freedom                               780         780
#>   P-value                                        0.000       0.000
#>   Scaling correction factor                                  7.195
#> 
#> User Model versus Baseline Model:
#> 
#>   Comparative Fit Index (CFI)                    0.995       0.951
#>   Tucker-Lewis Index (TLI)                       0.994       0.948
#>                                                                   
#>   Robust Comparative Fit Index (CFI)                            NA
#>   Robust Tucker-Lewis Index (TLI)                               NA
#> 
#> Root Mean Square Error of Approximation:
#> 
#>   RMSEA                                          0.066       0.074
#>   90 Percent confidence interval - lower         0.063       0.072
#>   90 Percent confidence interval - upper         0.068       0.076
#>   P-value RMSEA <= 0.05                          0.000       0.000
#>                                                                   
#>   Robust RMSEA                                                  NA
#>   90 Percent confidence interval - lower                        NA
#>   90 Percent confidence interval - upper                        NA
#> 
#> Standardized Root Mean Square Residual:
#> 
#>   SRMR                                           0.054       0.054

# }