Run Calibration

runCalibration is a function to perform item calibration on the response data.

runCalibration(
  data,
  dimensions = 1,
  fix_method = "free",
  fixedpar = NULL,
  ignore_nonconv = FALSE,
  ...
)

Arguments

data
a PROsetta_data object. See loadData for loading a dataset.

dimensions
number of dimensions to use. Must be 1 or 2. If 1, use one underlying dimension for all instruments combined. If 2, use each dimension separately for the anchor instrument and the developing instrument. Covariance between dimensions is freely estimated. (default = 1)

fix_method
the type of constraints to impose. (default = free)

item for fixed parameter calibration using anchor item parameters

theta for using the mean and the variance obtained from a unidimensional calibration of anchor items

free for free calibration

fixedpar
this argument exists for reproducibility. TRUE is equivalent to fix_method = "item", and FALSE is equivalent to fix_method = "free".

ignore_nonconv
if TRUE, return results even when calibration does not converge. If FALSE, raise an error when calibration does not converge. (default = FALSE)

...
additional arguments to pass onto mirt in 'mirt' package.

Value

runCalibration returns a SingleGroupClass object containing item calibration results.

This object can be used in coef, itemfit, itemplot in 'mirt' package to extract wanted information.

Examples

if (FALSE) {
out_calib <- runCalibration(data_asq) # errors
}
# \donttest{
out_calib <- runCalibration(data_asq, technical = list(NCYCLES = 1000))

mirt::coef(out_calib, IRTpars = TRUE, simplify = TRUE)
mirt::itemfit(out_calib, empirical.plot = 1)
mirt::itemplot(out_calib, item = 1, type = "info")
mirt::itemfit(out_calib, "S_X2", na.rm = TRUE)
# }