Run Test Equating

runEquateObserved is a function to perform equipercentile test equating between two scales. A concordance table is produced, mapping the observed raw scores from one scale to the scores from another scale.

runEquateObserved(
  data,
  scale_from = 2,
  scale_to = 1,
  type_to = "raw",
  rsss = NULL,
  eq_type = "equipercentile",
  smooth = "loglinear",
  degrees = list(3, 1),
  boot = TRUE,
  reps = 100,
  ...
)

Arguments

data
a PROsetta_data object. See loadData for loading a dataset.

scale_from
the scale ID of the input scale. References to itemmap in data argument. (default = 2)

scale_to
the scale ID of the target scale to equate to. References to itemmap in data argument. (default = 1)

type_to
the type of score to use in the target scale frequency table. Accepts raw, tscore, and theta. tscore and theta require argument rsss to be supplied. (default = raw)

rsss
the RSSS table to use to map each raw score level onto a t-score or a theta. See runRSSS.

eq_type
the type of equating to be passed onto equate in 'equate' package. (default = equipercentile)

smooth
the type of smoothing method to be passed onto presmoothing in 'equate' package. (default = loglinear)

degrees
the degrees of smoothing to be passed onto presmoothing. (default = list(3, 1))

boot
performs bootstrapping if TRUE. (default = TRUE)

reps
the number of replications to perform in bootstrapping. (default = 100)

...
other arguments to pass onto equate.

Value

runEquateObserved returns an equate object containing the test equating result.

The printed summary statistics indicate the distributional properties of the two supplied scales and the equated scale.

x corresponds to scale_from.

y corresponds to scale_to.

yx corresponds to scale_from after equating to scale_to.

See equate for details.

The concordance table is stored in concordance slot.

Examples

out_eq_raw <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  eq_type = "equipercentile", smooth = "loglinear"
)

#> runEquateObserved requires complete data, attempting to remove cases

#> getCompleteData: filtered 8 cases with missing responses in all scales

#> performing loglinear presmoothing on scale 2 (scale_from) distribution

#> performing loglinear presmoothing on scale 1 (scale_to) distribution

out_eq_raw$concordance

#>    raw_2     raw_1   raw_1_se raw_1_se_boot
#> 1     11  29.71858  0.2483603     0.1776580
#> 2     12  32.33361  0.4663568     0.4567534
#> 3     13  35.10463  0.6681376     0.6431408
#> 4     14  38.02455  0.8662512     0.8745018
#> 5     15  41.09143  1.0609494     1.1151137
#> 6     16  44.29367  1.2524657     1.2376782
#> 7     17  47.61721  1.5077675     1.4708558
#> 8     18  51.04427  1.6805879     1.7977498
#> 9     19  54.53510  1.9308974     2.0485023
#> 10    20  58.07716  2.0793550     2.2498398
#> 11    21  61.62444  2.3104050     2.4505889
#> 12    22  65.16029  2.4379903     2.7390007
#> 13    23  68.65089  2.6546378     2.8919124
#> 14    24  72.08339  2.7663267     3.0725263
#> 15    25  75.43304  2.8846186     3.1458178
#> 16    26  78.69900  3.1006498     3.1075550
#> 17    27  81.86936  3.2176468     3.2448089
#> 18    28  84.94022  3.3460076     3.2878864
#> 19    29  87.91190  3.4893302     3.2798016
#> 20    30  90.78567  3.6547336     3.5960570
#> 21    31  93.56298  3.8541951     4.1598912
#> 22    32  96.25648  3.8828176     4.2439661
#> 23    33  98.86327  4.1245578     4.2912381
#> 24    34 101.38572  4.2016116     4.4460155
#> 25    35 103.84034  4.5176063     4.5052509
#> 26    36 106.22205  4.6330766     4.7796254
#> 27    37 108.53555  5.0946338     5.1840713
#> 28    38 110.80177  5.2584873     5.7175696
#> 29    39 113.01125  5.4650506     6.5375165
#> 30    40 115.17125  5.7131451     6.7778803
#> 31    41 117.28871  6.0029191     7.1314362
#> 32    42 119.37049  6.3357393     7.5557551
#> 33    43 121.42353  6.7139238     8.2436651
#> 34    44 123.45521  7.1402725     9.4558585
#> 35    45 125.47375  7.6172335     9.9570781
#> 36    46 127.48883  8.1454867    10.0979919
#> 37    47 129.51373  9.6756492    10.9796842
#> 38    48 131.56670 10.3680795    10.9644244
#> 39    49 133.67002 11.0295096     9.7580908
#> 40    50 135.85632 11.5574281     9.5228130
#> 41    51 138.17184 11.7867191     8.5194702
#> 42    52 140.70935 12.8142324     7.0466973
#> 43    53 143.65612 10.7336688     4.4570932

# \donttest{
out_link <- runLinking(data_asq, method = "FIXEDPAR")

#> response data has 40 items

#> performing 1D fixed parameter calibration, using anchor data

#> anchor has 29 items * 5 parameters = 145 parameters

#> anchor parameters applied to par_layout

#> Iteration: 1, Log-Lik: -24053.720, Max-Change: 2.73844Iteration: 2, Log-Lik: -23453.207, Max-Change: 0.72611Iteration: 3, Log-Lik: -23439.758, Max-Change: 0.41693Iteration: 4, Log-Lik: -23438.315, Max-Change: 0.66336Iteration: 5, Log-Lik: -23437.791, Max-Change: 0.50564Iteration: 6, Log-Lik: -23437.683, Max-Change: 0.13194Iteration: 7, Log-Lik: -23437.680, Max-Change: 0.08704Iteration: 8, Log-Lik: -23437.677, Max-Change: 0.00500Iteration: 9, Log-Lik: -23437.677, Max-Change: 0.00021Iteration: 10, Log-Lik: -23437.677, Max-Change: 0.00014Iteration: 11, Log-Lik: -23437.677, Max-Change: 0.00014Iteration: 12, Log-Lik: -23437.677, Max-Change: 0.00005

out_rsss <- runRSSS(data_asq, out_link)
out_eq_tscore <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  type_to = "tscore", rsss = out_rsss,
  eq_type = "equipercentile", smooth = "loglinear"
)

#> runEquateObserved requires complete data, attempting to remove cases

#> getCompleteData: filtered 8 cases with missing responses in all scales

#> performing loglinear presmoothing on scale 2 (scale_from) distribution

#> mapping scale 1 (scale_to) raw scores to t-scores using supplied rsss

#> performing loglinear presmoothing on scale 1 (scale_to) distribution

out_eq_tscore$concordance

#>    raw_2 tscore_1 tscore_1_se tscore_1_se_boot
#> 1     11 33.36372   0.1840702        0.6458357
#> 2     12 39.16756   0.4814386        0.6057601
#> 3     13 42.67555   0.7317634        0.7372714
#> 4     14 45.26483   0.9456388        0.6848983
#> 5     15 47.32618   1.1381595        0.6784953
#> 6     16 49.14026   1.3681523        0.7061719
#> 7     17 50.67257   1.5144268        0.7159583
#> 8     18 52.06482   1.7168981        0.7647180
#> 9     19 53.42663   1.8602303        0.7431642
#> 10    20 54.63536   2.0595410        0.7251766
#> 11    21 55.80750   2.1741094        0.7235577
#> 12    22 56.91734   2.3016332        0.7324985
#> 13    23 58.01375   2.5128914        0.7591992
#> 14    24 59.09321   2.6519517        0.8509649
#> 15    25 60.13984   2.8044582        0.8998441
#> 16    26 61.10221   2.9328283        1.0100719
#> 17    27 62.13172   3.1087742        1.1590277
#> 18    28 63.09020   3.2540073        1.2353131
#> 19    29 64.03928   3.4548135        1.2719089
#> 20    30 65.02082   3.6725052        1.3661158
#> 21    31 65.92092   3.8530508        1.3627815
#> 22    32 66.82951   4.1109814        1.3811495
#> 23    33 67.75739   4.4021366        1.5352137
#> 24    34 68.66886   4.7385995        1.6928432
#> 25    35 69.49840   4.7638406        1.8093149
#> 26    36 70.35314   5.1299518        1.8385719
#> 27    37 71.19414   5.2664352        1.8873500
#> 28    38 72.02934   5.6878068        1.9958370
#> 29    39 72.82289   5.8343796        2.0989295
#> 30    40 73.62732   6.0120004        2.0727823
#> 31    41 74.40785   6.5136427        2.6575850
#> 32    42 75.21162   6.8017115        2.6908314
#> 33    43 75.95872   6.9701691        2.6761068
#> 34    44 76.76154   7.2782573        3.1035202
#> 35    45 77.53826   7.5895778        4.1768730
#> 36    46 78.34769   8.0697432        4.4810275
#> 37    47 79.17070   8.5724514        5.2203110
#> 38    48 80.07361   8.2175842        5.3895602
#> 39    49 81.09158   8.8627216        5.2838622
#> 40    50 82.21165   9.6443142        4.9467569
#> 41    51 83.52426  10.4101281        4.2961324
#> 42    52 85.31367  11.3165392        3.6374794
#> 43    53 87.50312  11.0429480        2.3775803

# }