Run Scale Linking

runLinking is a function to obtain item parameters from the response data, and perform scale linking onto the metric of supplied anchor item parameters.

runLinking(data, method, ...)

Arguments

data
a PROsetta_data object. See loadData for loading a dataset.

method
the type of linking to perform. Accepts:

MM for mean-mean

MS for mean-sigma

HB for Haebara method

SL for Stocking-Lord method

FIXEDPAR for fixed parameter calibration

Linear transformation methods are performed with plink in 'plink' package.

...
additional arguments to pass onto mirt in 'mirt' package.

Value

runLinking returns a list containing the scale linking results.

constants linear transformation constants. NA if method argument was FIXEDPAR.

ipar_linked item parameters calibrated to the response data, and linked to the anchor item parameters.

ipar_anchor anchor item parameters used in linking.

Examples

# \donttest{
out_link <- runLinking(data_asq, "SL", technical = list(NCYCLES = 1000))

#> response data has 40 items

#> performing free calibration of all items, ignoring anchor data

#> Iteration: 1, Log-Lik: -25455.892, Max-Change: 5.65432Iteration: 2, Log-Lik: -23797.560, Max-Change: 1.43832Iteration: 3, Log-Lik: -23628.029, Max-Change: 0.64961Iteration: 4, Log-Lik: -23548.749, Max-Change: 0.38055Iteration: 5, Log-Lik: -23500.707, Max-Change: 0.26763Iteration: 6, Log-Lik: -23461.760, Max-Change: 0.24313Iteration: 7, Log-Lik: -23431.938, Max-Change: 0.15656Iteration: 8, Log-Lik: -23407.322, Max-Change: 0.14222Iteration: 9, Log-Lik: -23388.165, Max-Change: 0.15375Iteration: 10, Log-Lik: -23371.592, Max-Change: 0.10763Iteration: 11, Log-Lik: -23353.037, Max-Change: 0.08862Iteration: 12, Log-Lik: -23339.073, Max-Change: 0.08080Iteration: 13, Log-Lik: -23328.299, Max-Change: 0.08490Iteration: 14, Log-Lik: -23320.188, Max-Change: 0.08289Iteration: 15, Log-Lik: -23315.310, Max-Change: 0.05720Iteration: 16, Log-Lik: -23309.518, Max-Change: 0.07428Iteration: 17, Log-Lik: -23305.336, Max-Change: 0.08033Iteration: 18, Log-Lik: -23301.939, Max-Change: 0.10009Iteration: 19, Log-Lik: -23299.170, Max-Change: 0.03978Iteration: 20, Log-Lik: -23296.063, Max-Change: 0.04314Iteration: 21, Log-Lik: -23293.461, Max-Change: 0.03392Iteration: 22, Log-Lik: -23288.875, Max-Change: 0.04563Iteration: 23, Log-Lik: -23287.114, Max-Change: 0.03085Iteration: 24, Log-Lik: -23285.620, Max-Change: 0.02825Iteration: 25, Log-Lik: -23282.533, Max-Change: 0.02388Iteration: 26, Log-Lik: -23281.576, Max-Change: 0.01945Iteration: 27, Log-Lik: -23280.889, Max-Change: 0.01837Iteration: 28, Log-Lik: -23278.377, Max-Change: 0.01704Iteration: 29, Log-Lik: -23277.857, Max-Change: 0.01216Iteration: 30, Log-Lik: -23277.413, Max-Change: 0.01177Iteration: 31, Log-Lik: -23276.097, Max-Change: 0.00948Iteration: 32, Log-Lik: -23275.771, Max-Change: 0.00947Iteration: 33, Log-Lik: -23275.492, Max-Change: 0.00965Iteration: 34, Log-Lik: -23274.384, Max-Change: 0.01421Iteration: 35, Log-Lik: -23274.112, Max-Change: 0.00786Iteration: 36, Log-Lik: -23273.905, Max-Change: 0.00612Iteration: 37, Log-Lik: -23273.567, Max-Change: 0.00702Iteration: 38, Log-Lik: -23273.389, Max-Change: 0.00603Iteration: 39, Log-Lik: -23273.216, Max-Change: 0.00646Iteration: 40, Log-Lik: -23272.995, Max-Change: 0.00532Iteration: 41, Log-Lik: -23272.836, Max-Change: 0.00733Iteration: 42, Log-Lik: -23272.687, Max-Change: 0.00577Iteration: 43, Log-Lik: -23272.537, Max-Change: 0.00504Iteration: 44, Log-Lik: -23272.395, Max-Change: 0.00790Iteration: 45, Log-Lik: -23272.267, Max-Change: 0.00543Iteration: 46, Log-Lik: -23272.169, Max-Change: 0.00666Iteration: 47, Log-Lik: -23272.054, Max-Change: 0.00500Iteration: 48, Log-Lik: -23271.944, Max-Change: 0.00485Iteration: 49, Log-Lik: -23271.714, Max-Change: 0.00555Iteration: 50, Log-Lik: -23271.604, Max-Change: 0.00763Iteration: 51, Log-Lik: -23271.509, Max-Change: 0.00399Iteration: 52, Log-Lik: -23271.395, Max-Change: 0.00610Iteration: 53, Log-Lik: -23271.306, Max-Change: 0.00394Iteration: 54, Log-Lik: -23271.221, Max-Change: 0.00596Iteration: 55, Log-Lik: -23271.060, Max-Change: 0.01442Iteration: 56, Log-Lik: -23270.981, Max-Change: 0.00444Iteration: 57, Log-Lik: -23270.905, Max-Change: 0.00377Iteration: 58, Log-Lik: -23270.724, Max-Change: 0.00933Iteration: 59, Log-Lik: -23270.658, Max-Change: 0.00397Iteration: 60, Log-Lik: -23270.591, Max-Change: 0.00536Iteration: 61, Log-Lik: -23270.482, Max-Change: 0.00373Iteration: 62, Log-Lik: -23270.429, Max-Change: 0.00521Iteration: 63, Log-Lik: -23270.374, Max-Change: 0.00336Iteration: 64, Log-Lik: -23270.311, Max-Change: 0.00504Iteration: 65, Log-Lik: -23270.261, Max-Change: 0.00344Iteration: 66, Log-Lik: -23270.216, Max-Change: 0.00612Iteration: 67, Log-Lik: -23270.153, Max-Change: 0.00414Iteration: 68, Log-Lik: -23270.108, Max-Change: 0.00573Iteration: 69, Log-Lik: -23270.063, Max-Change: 0.00309Iteration: 70, Log-Lik: -23270.028, Max-Change: 0.00523Iteration: 71, Log-Lik: -23269.987, Max-Change: 0.00318Iteration: 72, Log-Lik: -23269.950, Max-Change: 0.00470Iteration: 73, Log-Lik: -23269.897, Max-Change: 0.00314Iteration: 74, Log-Lik: -23269.862, Max-Change: 0.00352Iteration: 75, Log-Lik: -23269.828, Max-Change: 0.00328Iteration: 76, Log-Lik: -23269.787, Max-Change: 0.00625Iteration: 77, Log-Lik: -23269.752, Max-Change: 0.00298Iteration: 78, Log-Lik: -23269.722, Max-Change: 0.00583Iteration: 79, Log-Lik: -23269.681, Max-Change: 0.00284Iteration: 80, Log-Lik: -23269.653, Max-Change: 0.00278Iteration: 81, Log-Lik: -23269.626, Max-Change: 0.00286Iteration: 82, Log-Lik: -23269.576, Max-Change: 0.00660Iteration: 83, Log-Lik: -23269.547, Max-Change: 0.00262Iteration: 84, Log-Lik: -23269.523, Max-Change: 0.00541Iteration: 85, Log-Lik: -23269.490, Max-Change: 0.00270Iteration: 86, Log-Lik: -23269.468, Max-Change: 0.00261Iteration: 87, Log-Lik: -23269.446, Max-Change: 0.00259Iteration: 88, Log-Lik: -23269.421, Max-Change: 0.00587Iteration: 89, Log-Lik: -23269.398, Max-Change: 0.00258Iteration: 90, Log-Lik: -23269.378, Max-Change: 0.00502Iteration: 91, Log-Lik: -23269.351, Max-Change: 0.00230Iteration: 92, Log-Lik: -23269.333, Max-Change: 0.00222Iteration: 93, Log-Lik: -23269.316, Max-Change: 0.00211Iteration: 94, Log-Lik: -23269.291, Max-Change: 0.00524Iteration: 95, Log-Lik: -23269.271, Max-Change: 0.00214Iteration: 96, Log-Lik: -23269.257, Max-Change: 0.00493Iteration: 97, Log-Lik: -23269.235, Max-Change: 0.00216Iteration: 98, Log-Lik: -23269.221, Max-Change: 0.00190Iteration: 99, Log-Lik: -23269.208, Max-Change: 0.00303Iteration: 100, Log-Lik: -23269.186, Max-Change: 0.00264Iteration: 101, Log-Lik: -23269.173, Max-Change: 0.00172Iteration: 102, Log-Lik: -23269.162, Max-Change: 0.00500Iteration: 103, Log-Lik: -23269.144, Max-Change: 0.00212Iteration: 104, Log-Lik: -23269.132, Max-Change: 0.00173Iteration: 105, Log-Lik: -23269.121, Max-Change: 0.00234Iteration: 106, Log-Lik: -23269.101, Max-Change: 0.00180Iteration: 107, Log-Lik: -23269.091, Max-Change: 0.00167Iteration: 108, Log-Lik: -23269.081, Max-Change: 0.00313Iteration: 109, Log-Lik: -23269.063, Max-Change: 0.00424Iteration: 110, Log-Lik: -23269.049, Max-Change: 0.00228Iteration: 111, Log-Lik: -23269.040, Max-Change: 0.00144Iteration: 112, Log-Lik: -23269.028, Max-Change: 0.00291Iteration: 113, Log-Lik: -23269.018, Max-Change: 0.00714Iteration: 114, Log-Lik: -23269.006, Max-Change: 0.00140Iteration: 115, Log-Lik: -23269.005, Max-Change: 0.00137Iteration: 116, Log-Lik: -23268.997, Max-Change: 0.00129Iteration: 117, Log-Lik: -23268.990, Max-Change: 0.00130Iteration: 118, Log-Lik: -23268.952, Max-Change: 0.00716Iteration: 119, Log-Lik: -23268.940, Max-Change: 0.00120Iteration: 120, Log-Lik: -23268.934, Max-Change: 0.00123Iteration: 121, Log-Lik: -23268.909, Max-Change: 0.00692Iteration: 122, Log-Lik: -23268.900, Max-Change: 0.00148Iteration: 123, Log-Lik: -23268.895, Max-Change: 0.00100Iteration: 124, Log-Lik: -23268.886, Max-Change: 0.00184Iteration: 125, Log-Lik: -23268.882, Max-Change: 0.00112Iteration: 126, Log-Lik: -23268.878, Max-Change: 0.00103Iteration: 127, Log-Lik: -23268.871, Max-Change: 0.00108Iteration: 128, Log-Lik: -23268.868, Max-Change: 0.00083Iteration: 129, Log-Lik: -23268.864, Max-Change: 0.00084Iteration: 130, Log-Lik: -23268.847, Max-Change: 0.00776Iteration: 131, Log-Lik: -23268.838, Max-Change: 0.00112Iteration: 132, Log-Lik: -23268.835, Max-Change: 0.00102Iteration: 133, Log-Lik: -23268.829, Max-Change: 0.00183Iteration: 134, Log-Lik: -23268.826, Max-Change: 0.00173Iteration: 135, Log-Lik: -23268.823, Max-Change: 0.00080Iteration: 136, Log-Lik: -23268.821, Max-Change: 0.00065Iteration: 137, Log-Lik: -23268.819, Max-Change: 0.00067Iteration: 138, Log-Lik: -23268.817, Max-Change: 0.00066Iteration: 139, Log-Lik: -23268.807, Max-Change: 0.00062Iteration: 140, Log-Lik: -23268.805, Max-Change: 0.00057Iteration: 141, Log-Lik: -23268.804, Max-Change: 0.00060Iteration: 142, Log-Lik: -23268.799, Max-Change: 0.00133Iteration: 143, Log-Lik: -23268.798, Max-Change: 0.00061Iteration: 144, Log-Lik: -23268.796, Max-Change: 0.00595Iteration: 145, Log-Lik: -23268.791, Max-Change: 0.00155Iteration: 146, Log-Lik: -23268.789, Max-Change: 0.00083Iteration: 147, Log-Lik: -23268.788, Max-Change: 0.00064Iteration: 148, Log-Lik: -23268.786, Max-Change: 0.00201Iteration: 149, Log-Lik: -23268.786, Max-Change: 0.00166Iteration: 150, Log-Lik: -23268.784, Max-Change: 0.00186Iteration: 151, Log-Lik: -23268.784, Max-Change: 0.00076Iteration: 152, Log-Lik: -23268.784, Max-Change: 0.00137Iteration: 153, Log-Lik: -23268.783, Max-Change: 0.00184Iteration: 154, Log-Lik: -23268.783, Max-Change: 0.00183Iteration: 155, Log-Lik: -23268.782, Max-Change: 0.00051Iteration: 156, Log-Lik: -23268.782, Max-Change: 0.00122Iteration: 157, Log-Lik: -23268.782, Max-Change: 0.00234Iteration: 158, Log-Lik: -23268.781, Max-Change: 0.00212Iteration: 159, Log-Lik: -23268.781, Max-Change: 0.00105Iteration: 160, Log-Lik: -23268.778, Max-Change: 0.00175Iteration: 161, Log-Lik: -23268.778, Max-Change: 0.00050Iteration: 162, Log-Lik: -23268.778, Max-Change: 0.00120Iteration: 163, Log-Lik: -23268.778, Max-Change: 0.00100Iteration: 164, Log-Lik: -23268.776, Max-Change: 0.00176Iteration: 165, Log-Lik: -23268.776, Max-Change: 0.00050Iteration: 166, Log-Lik: -23268.776, Max-Change: 0.00044Iteration: 167, Log-Lik: -23268.775, Max-Change: 0.00106Iteration: 168, Log-Lik: -23268.775, Max-Change: 0.00175Iteration: 169, Log-Lik: -23268.775, Max-Change: 0.00138Iteration: 170, Log-Lik: -23268.775, Max-Change: 0.00045Iteration: 171, Log-Lik: -23268.774, Max-Change: 0.00115Iteration: 172, Log-Lik: -23268.774, Max-Change: 0.00053Iteration: 173, Log-Lik: -23268.774, Max-Change: 0.00129Iteration: 174, Log-Lik: -23268.774, Max-Change: 0.00209Iteration: 175, Log-Lik: -23268.774, Max-Change: 0.00145Iteration: 176, Log-Lik: -23268.773, Max-Change: 0.00047Iteration: 177, Log-Lik: -23268.773, Max-Change: 0.00120Iteration: 178, Log-Lik: -23268.773, Max-Change: 0.00056Iteration: 179, Log-Lik: -23268.773, Max-Change: 0.00137Iteration: 180, Log-Lik: -23268.773, Max-Change: 0.00044Iteration: 181, Log-Lik: -23268.773, Max-Change: 0.00196Iteration: 182, Log-Lik: -23268.772, Max-Change: 0.00056Iteration: 183, Log-Lik: -23268.772, Max-Change: 0.00136Iteration: 184, Log-Lik: -23268.772, Max-Change: 0.00640Iteration: 185, Log-Lik: -23268.768, Max-Change: 0.00103Iteration: 186, Log-Lik: -23268.767, Max-Change: 0.00083Iteration: 187, Log-Lik: -23268.767, Max-Change: 0.00048Iteration: 188, Log-Lik: -23268.767, Max-Change: 0.00088Iteration: 189, Log-Lik: -23268.766, Max-Change: 0.00109Iteration: 190, Log-Lik: -23268.766, Max-Change: 0.00143Iteration: 191, Log-Lik: -23268.766, Max-Change: 0.00039Iteration: 192, Log-Lik: -23268.766, Max-Change: 0.00093Iteration: 193, Log-Lik: -23268.766, Max-Change: 0.00052Iteration: 194, Log-Lik: -23268.765, Max-Change: 0.00133Iteration: 195, Log-Lik: -23268.765, Max-Change: 0.00036Iteration: 196, Log-Lik: -23268.765, Max-Change: 0.00159Iteration: 197, Log-Lik: -23268.765, Max-Change: 0.00051Iteration: 198, Log-Lik: -23268.765, Max-Change: 0.00132Iteration: 199, Log-Lik: -23268.765, Max-Change: 0.00086Iteration: 200, Log-Lik: -23268.763, Max-Change: 0.00121Iteration: 201, Log-Lik: -23268.763, Max-Change: 0.00040Iteration: 202, Log-Lik: -23268.763, Max-Change: 0.00035Iteration: 203, Log-Lik: -23268.763, Max-Change: 0.00091Iteration: 204, Log-Lik: -23268.763, Max-Change: 0.00123Iteration: 205, Log-Lik: -23268.763, Max-Change: 0.00118Iteration: 206, Log-Lik: -23268.763, Max-Change: 0.00033Iteration: 207, Log-Lik: -23268.763, Max-Change: 0.00079Iteration: 208, Log-Lik: -23268.762, Max-Change: 0.00043Iteration: 209, Log-Lik: -23268.762, Max-Change: 0.00110Iteration: 210, Log-Lik: -23268.762, Max-Change: 0.00153Iteration: 211, Log-Lik: -23268.762, Max-Change: 0.00123Iteration: 212, Log-Lik: -23268.762, Max-Change: 0.00034Iteration: 213, Log-Lik: -23268.762, Max-Change: 0.00083Iteration: 214, Log-Lik: -23268.762, Max-Change: 0.00046Iteration: 215, Log-Lik: -23268.762, Max-Change: 0.00116Iteration: 216, Log-Lik: -23268.761, Max-Change: 0.00033Iteration: 217, Log-Lik: -23268.761, Max-Change: 0.00143Iteration: 218, Log-Lik: -23268.761, Max-Change: 0.00045Iteration: 219, Log-Lik: -23268.761, Max-Change: 0.00116Iteration: 220, Log-Lik: -23268.761, Max-Change: 0.00058Iteration: 221, Log-Lik: -23268.761, Max-Change: 0.00142Iteration: 222, Log-Lik: -23268.761, Max-Change: 0.00045Iteration: 223, Log-Lik: -23268.761, Max-Change: 0.00040Iteration: 224, Log-Lik: -23268.761, Max-Change: 0.00102Iteration: 225, Log-Lik: -23268.761, Max-Change: 0.00142Iteration: 226, Log-Lik: -23268.760, Max-Change: 0.00115Iteration: 227, Log-Lik: -23268.760, Max-Change: 0.00032Iteration: 228, Log-Lik: -23268.760, Max-Change: 0.00078Iteration: 229, Log-Lik: -23268.760, Max-Change: 0.00043Iteration: 230, Log-Lik: -23268.760, Max-Change: 0.00108Iteration: 231, Log-Lik: -23268.760, Max-Change: 0.00030Iteration: 232, Log-Lik: -23268.760, Max-Change: 0.00134Iteration: 233, Log-Lik: -23268.760, Max-Change: 0.00042Iteration: 234, Log-Lik: -23268.760, Max-Change: 0.00108Iteration: 235, Log-Lik: -23268.760, Max-Change: 0.00054Iteration: 236, Log-Lik: -23268.759, Max-Change: 0.00133Iteration: 237, Log-Lik: -23268.759, Max-Change: 0.00042Iteration: 238, Log-Lik: -23268.759, Max-Change: 0.00037Iteration: 239, Log-Lik: -23268.759, Max-Change: 0.00095Iteration: 240, Log-Lik: -23268.759, Max-Change: 0.00133Iteration: 241, Log-Lik: -23268.759, Max-Change: 0.00108Iteration: 242, Log-Lik: -23268.759, Max-Change: 0.00030Iteration: 243, Log-Lik: -23268.759, Max-Change: 0.00074Iteration: 244, Log-Lik: -23268.759, Max-Change: 0.00040Iteration: 245, Log-Lik: -23268.759, Max-Change: 0.00102Iteration: 246, Log-Lik: -23268.759, Max-Change: 0.00143Iteration: 247, Log-Lik: -23268.758, Max-Change: 0.00109Iteration: 248, Log-Lik: -23268.758, Max-Change: 0.00031Iteration: 249, Log-Lik: -23268.758, Max-Change: 0.00075Iteration: 250, Log-Lik: -23268.758, Max-Change: 0.00041Iteration: 251, Log-Lik: -23268.758, Max-Change: 0.00104Iteration: 252, Log-Lik: -23268.758, Max-Change: 0.00029Iteration: 253, Log-Lik: -23268.758, Max-Change: 0.00129Iteration: 254, Log-Lik: -23268.758, Max-Change: 0.00041Iteration: 255, Log-Lik: -23268.758, Max-Change: 0.00103Iteration: 256, Log-Lik: -23268.758, Max-Change: 0.00052Iteration: 257, Log-Lik: -23268.758, Max-Change: 0.00129Iteration: 258, Log-Lik: -23268.758, Max-Change: 0.00041Iteration: 259, Log-Lik: -23268.758, Max-Change: 0.00036Iteration: 260, Log-Lik: -23268.757, Max-Change: 0.00092Iteration: 261, Log-Lik: -23268.757, Max-Change: 0.00129Iteration: 262, Log-Lik: -23268.757, Max-Change: 0.00101Iteration: 263, Log-Lik: -23268.757, Max-Change: 0.00029Iteration: 264, Log-Lik: -23268.757, Max-Change: 0.00070Iteration: 265, Log-Lik: -23268.757, Max-Change: 0.00038Iteration: 266, Log-Lik: -23268.757, Max-Change: 0.00096Iteration: 267, Log-Lik: -23268.757, Max-Change: 0.00136Iteration: 268, Log-Lik: -23268.757, Max-Change: 0.00102Iteration: 269, Log-Lik: -23268.757, Max-Change: 0.00029Iteration: 270, Log-Lik: -23268.757, Max-Change: 0.00071Iteration: 271, Log-Lik: -23268.757, Max-Change: 0.00038Iteration: 272, Log-Lik: -23268.756, Max-Change: 0.00098Iteration: 273, Log-Lik: -23268.756, Max-Change: 0.00028Iteration: 274, Log-Lik: -23268.756, Max-Change: 0.00121Iteration: 275, Log-Lik: -23268.756, Max-Change: 0.00038Iteration: 276, Log-Lik: -23268.756, Max-Change: 0.00097Iteration: 277, Log-Lik: -23268.756, Max-Change: 0.00049Iteration: 278, Log-Lik: -23268.756, Max-Change: 0.00122Iteration: 279, Log-Lik: -23268.756, Max-Change: 0.00038Iteration: 280, Log-Lik: -23268.756, Max-Change: 0.00034Iteration: 281, Log-Lik: -23268.756, Max-Change: 0.00086Iteration: 282, Log-Lik: -23268.756, Max-Change: 0.00121Iteration: 283, Log-Lik: -23268.756, Max-Change: 0.00095Iteration: 284, Log-Lik: -23268.756, Max-Change: 0.00027Iteration: 285, Log-Lik: -23268.756, Max-Change: 0.00066Iteration: 286, Log-Lik: -23268.756, Max-Change: 0.00035Iteration: 287, Log-Lik: -23268.755, Max-Change: 0.00090Iteration: 288, Log-Lik: -23268.755, Max-Change: 0.00128Iteration: 289, Log-Lik: -23268.755, Max-Change: 0.00096Iteration: 290, Log-Lik: -23268.755, Max-Change: 0.00027Iteration: 291, Log-Lik: -23268.755, Max-Change: 0.00066Iteration: 292, Log-Lik: -23268.755, Max-Change: 0.00036Iteration: 293, Log-Lik: -23268.755, Max-Change: 0.00091Iteration: 294, Log-Lik: -23268.755, Max-Change: 0.00026Iteration: 295, Log-Lik: -23268.755, Max-Change: 0.00114Iteration: 296, Log-Lik: -23268.755, Max-Change: 0.00036Iteration: 297, Log-Lik: -23268.755, Max-Change: 0.00091Iteration: 298, Log-Lik: -23268.755, Max-Change: 0.00046Iteration: 299, Log-Lik: -23268.755, Max-Change: 0.00114Iteration: 300, Log-Lik: -23268.755, Max-Change: 0.00036Iteration: 301, Log-Lik: -23268.755, Max-Change: 0.00032Iteration: 302, Log-Lik: -23268.754, Max-Change: 0.00081Iteration: 303, Log-Lik: -23268.754, Max-Change: 0.00114Iteration: 304, Log-Lik: -23268.754, Max-Change: 0.00089Iteration: 305, Log-Lik: -23268.754, Max-Change: 0.00025Iteration: 306, Log-Lik: -23268.754, Max-Change: 0.00062Iteration: 307, Log-Lik: -23268.754, Max-Change: 0.00033Iteration: 308, Log-Lik: -23268.754, Max-Change: 0.00085Iteration: 309, Log-Lik: -23268.754, Max-Change: 0.00120Iteration: 310, Log-Lik: -23268.754, Max-Change: 0.00090Iteration: 311, Log-Lik: -23268.754, Max-Change: 0.00026Iteration: 312, Log-Lik: -23268.754, Max-Change: 0.00062Iteration: 313, Log-Lik: -23268.754, Max-Change: 0.00034Iteration: 314, Log-Lik: -23268.754, Max-Change: 0.00086Iteration: 315, Log-Lik: -23268.754, Max-Change: 0.00024Iteration: 316, Log-Lik: -23268.754, Max-Change: 0.00107Iteration: 317, Log-Lik: -23268.754, Max-Change: 0.00034Iteration: 318, Log-Lik: -23268.753, Max-Change: 0.00085Iteration: 319, Log-Lik: -23268.754, Max-Change: 0.00044Iteration: 320, Log-Lik: -23268.753, Max-Change: 0.00107Iteration: 321, Log-Lik: -23268.753, Max-Change: 0.00034Iteration: 322, Log-Lik: -23268.753, Max-Change: 0.00030Iteration: 323, Log-Lik: -23268.753, Max-Change: 0.00076Iteration: 324, Log-Lik: -23268.753, Max-Change: 0.00107Iteration: 325, Log-Lik: -23268.753, Max-Change: 0.00084Iteration: 326, Log-Lik: -23268.753, Max-Change: 0.00024Iteration: 327, Log-Lik: -23268.753, Max-Change: 0.00058Iteration: 328, Log-Lik: -23268.753, Max-Change: 0.00031Iteration: 329, Log-Lik: -23268.753, Max-Change: 0.00079Iteration: 330, Log-Lik: -23268.753, Max-Change: 0.00113Iteration: 331, Log-Lik: -23268.753, Max-Change: 0.00084Iteration: 332, Log-Lik: -23268.753, Max-Change: 0.00024Iteration: 333, Log-Lik: -23268.753, Max-Change: 0.00059Iteration: 334, Log-Lik: -23268.753, Max-Change: 0.00032Iteration: 335, Log-Lik: -23268.753, Max-Change: 0.00081Iteration: 336, Log-Lik: -23268.752, Max-Change: 0.00023Iteration: 337, Log-Lik: -23268.752, Max-Change: 0.00101Iteration: 338, Log-Lik: -23268.752, Max-Change: 0.00032Iteration: 339, Log-Lik: -23268.752, Max-Change: 0.00080Iteration: 340, Log-Lik: -23268.752, Max-Change: 0.00041Iteration: 341, Log-Lik: -23268.752, Max-Change: 0.00101Iteration: 342, Log-Lik: -23268.752, Max-Change: 0.00032Iteration: 343, Log-Lik: -23268.752, Max-Change: 0.00028Iteration: 344, Log-Lik: -23268.752, Max-Change: 0.00071Iteration: 345, Log-Lik: -23268.752, Max-Change: 0.00101Iteration: 346, Log-Lik: -23268.752, Max-Change: 0.00079Iteration: 347, Log-Lik: -23268.752, Max-Change: 0.00022Iteration: 348, Log-Lik: -23268.752, Max-Change: 0.00055Iteration: 349, Log-Lik: -23268.752, Max-Change: 0.00029Iteration: 350, Log-Lik: -23268.752, Max-Change: 0.00075Iteration: 351, Log-Lik: -23268.752, Max-Change: 0.00106Iteration: 352, Log-Lik: -23268.752, Max-Change: 0.00079Iteration: 353, Log-Lik: -23268.752, Max-Change: 0.00023Iteration: 354, Log-Lik: -23268.752, Max-Change: 0.00055Iteration: 355, Log-Lik: -23268.752, Max-Change: 0.00030Iteration: 356, Log-Lik: -23268.751, Max-Change: 0.00076Iteration: 357, Log-Lik: -23268.751, Max-Change: 0.00022Iteration: 358, Log-Lik: -23268.751, Max-Change: 0.00095Iteration: 359, Log-Lik: -23268.751, Max-Change: 0.00030Iteration: 360, Log-Lik: -23268.751, Max-Change: 0.00075Iteration: 361, Log-Lik: -23268.751, Max-Change: 0.00039Iteration: 362, Log-Lik: -23268.751, Max-Change: 0.00095Iteration: 363, Log-Lik: -23268.751, Max-Change: 0.00030Iteration: 364, Log-Lik: -23268.751, Max-Change: 0.00026Iteration: 365, Log-Lik: -23268.751, Max-Change: 0.00067Iteration: 366, Log-Lik: -23268.751, Max-Change: 0.00095Iteration: 367, Log-Lik: -23268.751, Max-Change: 0.00074Iteration: 368, Log-Lik: -23268.751, Max-Change: 0.00021Iteration: 369, Log-Lik: -23268.751, Max-Change: 0.00051Iteration: 370, Log-Lik: -23268.751, Max-Change: 0.00028Iteration: 371, Log-Lik: -23268.751, Max-Change: 0.00070Iteration: 372, Log-Lik: -23268.751, Max-Change: 0.00100Iteration: 373, Log-Lik: -23268.751, Max-Change: 0.00075Iteration: 374, Log-Lik: -23268.751, Max-Change: 0.00021Iteration: 375, Log-Lik: -23268.751, Max-Change: 0.00052Iteration: 376, Log-Lik: -23268.751, Max-Change: 0.00028Iteration: 377, Log-Lik: -23268.751, Max-Change: 0.00071Iteration: 378, Log-Lik: -23268.751, Max-Change: 0.00020Iteration: 379, Log-Lik: -23268.751, Max-Change: 0.00089Iteration: 380, Log-Lik: -23268.751, Max-Change: 0.00028Iteration: 381, Log-Lik: -23268.750, Max-Change: 0.00071Iteration: 382, Log-Lik: -23268.750, Max-Change: 0.00036Iteration: 383, Log-Lik: -23268.750, Max-Change: 0.00089Iteration: 384, Log-Lik: -23268.750, Max-Change: 0.00028Iteration: 385, Log-Lik: -23268.750, Max-Change: 0.00025Iteration: 386, Log-Lik: -23268.750, Max-Change: 0.00063Iteration: 387, Log-Lik: -23268.750, Max-Change: 0.00089Iteration: 388, Log-Lik: -23268.750, Max-Change: 0.00070Iteration: 389, Log-Lik: -23268.750, Max-Change: 0.00020Iteration: 390, Log-Lik: -23268.750, Max-Change: 0.00048Iteration: 391, Log-Lik: -23268.750, Max-Change: 0.00026Iteration: 392, Log-Lik: -23268.750, Max-Change: 0.00066Iteration: 393, Log-Lik: -23268.750, Max-Change: 0.00094Iteration: 394, Log-Lik: -23268.750, Max-Change: 0.00070Iteration: 395, Log-Lik: -23268.750, Max-Change: 0.00020Iteration: 396, Log-Lik: -23268.750, Max-Change: 0.00049Iteration: 397, Log-Lik: -23268.750, Max-Change: 0.00026Iteration: 398, Log-Lik: -23268.750, Max-Change: 0.00067Iteration: 399, Log-Lik: -23268.750, Max-Change: 0.00019Iteration: 400, Log-Lik: -23268.750, Max-Change: 0.00084Iteration: 401, Log-Lik: -23268.750, Max-Change: 0.00026Iteration: 402, Log-Lik: -23268.750, Max-Change: 0.00067Iteration: 403, Log-Lik: -23268.750, Max-Change: 0.00034Iteration: 404, Log-Lik: -23268.750, Max-Change: 0.00084Iteration: 405, Log-Lik: -23268.750, Max-Change: 0.00026Iteration: 406, Log-Lik: -23268.750, Max-Change: 0.00023Iteration: 407, Log-Lik: -23268.750, Max-Change: 0.00059Iteration: 408, Log-Lik: -23268.749, Max-Change: 0.00084Iteration: 409, Log-Lik: -23268.749, Max-Change: 0.00066Iteration: 410, Log-Lik: -23268.749, Max-Change: 0.00019Iteration: 411, Log-Lik: -23268.749, Max-Change: 0.00046Iteration: 412, Log-Lik: -23268.749, Max-Change: 0.00025Iteration: 413, Log-Lik: -23268.749, Max-Change: 0.00062Iteration: 414, Log-Lik: -23268.749, Max-Change: 0.00089Iteration: 415, Log-Lik: -23268.749, Max-Change: 0.00066Iteration: 416, Log-Lik: -23268.749, Max-Change: 0.00019Iteration: 417, Log-Lik: -23268.749, Max-Change: 0.00046Iteration: 418, Log-Lik: -23268.749, Max-Change: 0.00025Iteration: 419, Log-Lik: -23268.749, Max-Change: 0.00063Iteration: 420, Log-Lik: -23268.749, Max-Change: 0.00018Iteration: 421, Log-Lik: -23268.749, Max-Change: 0.00079Iteration: 422, Log-Lik: -23268.749, Max-Change: 0.00025Iteration: 423, Log-Lik: -23268.749, Max-Change: 0.00063Iteration: 424, Log-Lik: -23268.749, Max-Change: 0.00032Iteration: 425, Log-Lik: -23268.749, Max-Change: 0.00079Iteration: 426, Log-Lik: -23268.749, Max-Change: 0.00025Iteration: 427, Log-Lik: -23268.749, Max-Change: 0.00022Iteration: 428, Log-Lik: -23268.749, Max-Change: 0.00056Iteration: 429, Log-Lik: -23268.749, Max-Change: 0.00079Iteration: 430, Log-Lik: -23268.749, Max-Change: 0.00062Iteration: 431, Log-Lik: -23268.749, Max-Change: 0.00018Iteration: 432, Log-Lik: -23268.749, Max-Change: 0.00043Iteration: 433, Log-Lik: -23268.749, Max-Change: 0.00023Iteration: 434, Log-Lik: -23268.749, Max-Change: 0.00059Iteration: 435, Log-Lik: -23268.749, Max-Change: 0.00084Iteration: 436, Log-Lik: -23268.749, Max-Change: 0.00063Iteration: 437, Log-Lik: -23268.749, Max-Change: 0.00018Iteration: 438, Log-Lik: -23268.749, Max-Change: 0.00044Iteration: 439, Log-Lik: -23268.748, Max-Change: 0.00024Iteration: 440, Log-Lik: -23268.748, Max-Change: 0.00060Iteration: 441, Log-Lik: -23268.748, Max-Change: 0.00017Iteration: 442, Log-Lik: -23268.748, Max-Change: 0.00075Iteration: 443, Log-Lik: -23268.748, Max-Change: 0.00023Iteration: 444, Log-Lik: -23268.748, Max-Change: 0.00059Iteration: 445, Log-Lik: -23268.748, Max-Change: 0.00030Iteration: 446, Log-Lik: -23268.748, Max-Change: 0.00075Iteration: 447, Log-Lik: -23268.748, Max-Change: 0.00023Iteration: 448, Log-Lik: -23268.748, Max-Change: 0.00021Iteration: 449, Log-Lik: -23268.748, Max-Change: 0.00053Iteration: 450, Log-Lik: -23268.748, Max-Change: 0.00075Iteration: 451, Log-Lik: -23268.748, Max-Change: 0.00058Iteration: 452, Log-Lik: -23268.748, Max-Change: 0.00017Iteration: 453, Log-Lik: -23268.748, Max-Change: 0.00041Iteration: 454, Log-Lik: -23268.748, Max-Change: 0.00022Iteration: 455, Log-Lik: -23268.748, Max-Change: 0.00055Iteration: 456, Log-Lik: -23268.748, Max-Change: 0.00079Iteration: 457, Log-Lik: -23268.748, Max-Change: 0.00059Iteration: 458, Log-Lik: -23268.748, Max-Change: 0.00017Iteration: 459, Log-Lik: -23268.748, Max-Change: 0.00041Iteration: 460, Log-Lik: -23268.748, Max-Change: 0.00022Iteration: 461, Log-Lik: -23268.748, Max-Change: 0.00056Iteration: 462, Log-Lik: -23268.748, Max-Change: 0.00016Iteration: 463, Log-Lik: -23268.748, Max-Change: 0.00071Iteration: 464, Log-Lik: -23268.748, Max-Change: 0.00022Iteration: 465, Log-Lik: -23268.748, Max-Change: 0.00056Iteration: 466, Log-Lik: -23268.748, Max-Change: 0.00029Iteration: 467, Log-Lik: -23268.748, Max-Change: 0.00071Iteration: 468, Log-Lik: -23268.748, Max-Change: 0.00022Iteration: 469, Log-Lik: -23268.748, Max-Change: 0.00020Iteration: 470, Log-Lik: -23268.748, Max-Change: 0.00050Iteration: 471, Log-Lik: -23268.748, Max-Change: 0.00070Iteration: 472, Log-Lik: -23268.748, Max-Change: 0.00055Iteration: 473, Log-Lik: -23268.748, Max-Change: 0.00016Iteration: 474, Log-Lik: -23268.748, Max-Change: 0.00038Iteration: 475, Log-Lik: -23268.748, Max-Change: 0.00021Iteration: 476, Log-Lik: -23268.748, Max-Change: 0.00052Iteration: 477, Log-Lik: -23268.748, Max-Change: 0.00074Iteration: 478, Log-Lik: -23268.747, Max-Change: 0.00056Iteration: 479, Log-Lik: -23268.747, Max-Change: 0.00016Iteration: 480, Log-Lik: -23268.747, Max-Change: 0.00039Iteration: 481, Log-Lik: -23268.747, Max-Change: 0.00021Iteration: 482, Log-Lik: -23268.747, Max-Change: 0.00053Iteration: 483, Log-Lik: -23268.747, Max-Change: 0.00015Iteration: 484, Log-Lik: -23268.747, Max-Change: 0.00067Iteration: 485, Log-Lik: -23268.747, Max-Change: 0.00021Iteration: 486, Log-Lik: -23268.747, Max-Change: 0.00053Iteration: 487, Log-Lik: -23268.747, Max-Change: 0.00027Iteration: 488, Log-Lik: -23268.747, Max-Change: 0.00067Iteration: 489, Log-Lik: -23268.747, Max-Change: 0.00021Iteration: 490, Log-Lik: -23268.747, Max-Change: 0.00018Iteration: 491, Log-Lik: -23268.747, Max-Change: 0.00047Iteration: 492, Log-Lik: -23268.747, Max-Change: 0.00066Iteration: 493, Log-Lik: -23268.747, Max-Change: 0.00052Iteration: 494, Log-Lik: -23268.747, Max-Change: 0.00015Iteration: 495, Log-Lik: -23268.747, Max-Change: 0.00036Iteration: 496, Log-Lik: -23268.747, Max-Change: 0.00019Iteration: 497, Log-Lik: -23268.747, Max-Change: 0.00049Iteration: 498, Log-Lik: -23268.747, Max-Change: 0.00070Iteration: 499, Log-Lik: -23268.747, Max-Change: 0.00053Iteration: 500, Log-Lik: -23268.747, Max-Change: 0.00015Iteration: 501, Log-Lik: -23268.747, Max-Change: 0.00037Iteration: 502, Log-Lik: -23268.747, Max-Change: 0.00020Iteration: 503, Log-Lik: -23268.747, Max-Change: 0.00050Iteration: 504, Log-Lik: -23268.747, Max-Change: 0.00014Iteration: 505, Log-Lik: -23268.747, Max-Change: 0.00063Iteration: 506, Log-Lik: -23268.747, Max-Change: 0.00020Iteration: 507, Log-Lik: -23268.747, Max-Change: 0.00050Iteration: 508, Log-Lik: -23268.747, Max-Change: 0.00026Iteration: 509, Log-Lik: -23268.747, Max-Change: 0.00063Iteration: 510, Log-Lik: -23268.747, Max-Change: 0.00020Iteration: 511, Log-Lik: -23268.747, Max-Change: 0.00017Iteration: 512, Log-Lik: -23268.747, Max-Change: 0.00044Iteration: 513, Log-Lik: -23268.747, Max-Change: 0.00063Iteration: 514, Log-Lik: -23268.747, Max-Change: 0.00049Iteration: 515, Log-Lik: -23268.747, Max-Change: 0.00014Iteration: 516, Log-Lik: -23268.747, Max-Change: 0.00034Iteration: 517, Log-Lik: -23268.747, Max-Change: 0.00018Iteration: 518, Log-Lik: -23268.747, Max-Change: 0.00046Iteration: 519, Log-Lik: -23268.747, Max-Change: 0.00066Iteration: 520, Log-Lik: -23268.747, Max-Change: 0.00050Iteration: 521, Log-Lik: -23268.747, Max-Change: 0.00014Iteration: 522, Log-Lik: -23268.747, Max-Change: 0.00035Iteration: 523, Log-Lik: -23268.747, Max-Change: 0.00019Iteration: 524, Log-Lik: -23268.747, Max-Change: 0.00047Iteration: 525, Log-Lik: -23268.747, Max-Change: 0.00013Iteration: 526, Log-Lik: -23268.747, Max-Change: 0.00059Iteration: 527, Log-Lik: -23268.747, Max-Change: 0.00019Iteration: 528, Log-Lik: -23268.746, Max-Change: 0.00047Iteration: 529, Log-Lik: -23268.746, Max-Change: 0.00024Iteration: 530, Log-Lik: -23268.746, Max-Change: 0.00059Iteration: 531, Log-Lik: -23268.746, Max-Change: 0.00019Iteration: 532, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 533, Log-Lik: -23268.746, Max-Change: 0.00042Iteration: 534, Log-Lik: -23268.746, Max-Change: 0.00059Iteration: 535, Log-Lik: -23268.746, Max-Change: 0.00046Iteration: 536, Log-Lik: -23268.746, Max-Change: 0.00013Iteration: 537, Log-Lik: -23268.746, Max-Change: 0.00032Iteration: 538, Log-Lik: -23268.746, Max-Change: 0.00017Iteration: 539, Log-Lik: -23268.746, Max-Change: 0.00044Iteration: 540, Log-Lik: -23268.746, Max-Change: 0.00062Iteration: 541, Log-Lik: -23268.746, Max-Change: 0.00047Iteration: 542, Log-Lik: -23268.746, Max-Change: 0.00013Iteration: 543, Log-Lik: -23268.746, Max-Change: 0.00033Iteration: 544, Log-Lik: -23268.746, Max-Change: 0.00018Iteration: 545, Log-Lik: -23268.746, Max-Change: 0.00045Iteration: 546, Log-Lik: -23268.746, Max-Change: 0.00064Iteration: 547, Log-Lik: -23268.746, Max-Change: 0.00047Iteration: 548, Log-Lik: -23268.746, Max-Change: 0.00013Iteration: 549, Log-Lik: -23268.746, Max-Change: 0.00033Iteration: 550, Log-Lik: -23268.746, Max-Change: 0.00018Iteration: 551, Log-Lik: -23268.746, Max-Change: 0.00045Iteration: 552, Log-Lik: -23268.746, Max-Change: 0.00013Iteration: 553, Log-Lik: -23268.746, Max-Change: 0.00056Iteration: 554, Log-Lik: -23268.746, Max-Change: 0.00018Iteration: 555, Log-Lik: -23268.746, Max-Change: 0.00045Iteration: 556, Log-Lik: -23268.746, Max-Change: 0.00023Iteration: 557, Log-Lik: -23268.746, Max-Change: 0.00056Iteration: 558, Log-Lik: -23268.746, Max-Change: 0.00018Iteration: 559, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 560, Log-Lik: -23268.746, Max-Change: 0.00040Iteration: 561, Log-Lik: -23268.746, Max-Change: 0.00056Iteration: 562, Log-Lik: -23268.746, Max-Change: 0.00043Iteration: 563, Log-Lik: -23268.746, Max-Change: 0.00012Iteration: 564, Log-Lik: -23268.746, Max-Change: 0.00030Iteration: 565, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 566, Log-Lik: -23268.746, Max-Change: 0.00041Iteration: 567, Log-Lik: -23268.746, Max-Change: 0.00059Iteration: 568, Log-Lik: -23268.746, Max-Change: 0.00044Iteration: 569, Log-Lik: -23268.746, Max-Change: 0.00012Iteration: 570, Log-Lik: -23268.746, Max-Change: 0.00031Iteration: 571, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 572, Log-Lik: -23268.746, Max-Change: 0.00042Iteration: 573, Log-Lik: -23268.746, Max-Change: 0.00012Iteration: 574, Log-Lik: -23268.746, Max-Change: 0.00052Iteration: 575, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 576, Log-Lik: -23268.746, Max-Change: 0.00041Iteration: 577, Log-Lik: -23268.746, Max-Change: 0.00021Iteration: 578, Log-Lik: -23268.746, Max-Change: 0.00052Iteration: 579, Log-Lik: -23268.746, Max-Change: 0.00016Iteration: 580, Log-Lik: -23268.746, Max-Change: 0.00014Iteration: 581, Log-Lik: -23268.746, Max-Change: 0.00037Iteration: 582, Log-Lik: -23268.746, Max-Change: 0.00052Iteration: 583, Log-Lik: -23268.746, Max-Change: 0.00041Iteration: 584, Log-Lik: -23268.746, Max-Change: 0.00012Iteration: 585, Log-Lik: -23268.746, Max-Change: 0.00028Iteration: 586, Log-Lik: -23268.746, Max-Change: 0.00015Iteration: 587, Log-Lik: -23268.746, Max-Change: 0.00039Iteration: 588, Log-Lik: -23268.746, Max-Change: 0.00055Iteration: 589, Log-Lik: -23268.746, Max-Change: 0.00041Iteration: 590, Log-Lik: -23268.746, Max-Change: 0.00012Iteration: 591, Log-Lik: -23268.746, Max-Change: 0.00029Iteration: 592, Log-Lik: -23268.746, Max-Change: 0.00015Iteration: 593, Log-Lik: -23268.745, Max-Change: 0.00039Iteration: 594, Log-Lik: -23268.745, Max-Change: 0.00011Iteration: 595, Log-Lik: -23268.745, Max-Change: 0.00049Iteration: 596, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 597, Log-Lik: -23268.745, Max-Change: 0.00039Iteration: 598, Log-Lik: -23268.745, Max-Change: 0.00020Iteration: 599, Log-Lik: -23268.745, Max-Change: 0.00049Iteration: 600, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 601, Log-Lik: -23268.745, Max-Change: 0.00014Iteration: 602, Log-Lik: -23268.745, Max-Change: 0.00035Iteration: 603, Log-Lik: -23268.745, Max-Change: 0.00049Iteration: 604, Log-Lik: -23268.745, Max-Change: 0.00038Iteration: 605, Log-Lik: -23268.745, Max-Change: 0.00011Iteration: 606, Log-Lik: -23268.745, Max-Change: 0.00027Iteration: 607, Log-Lik: -23268.745, Max-Change: 0.00014Iteration: 608, Log-Lik: -23268.745, Max-Change: 0.00036Iteration: 609, Log-Lik: -23268.745, Max-Change: 0.00052Iteration: 610, Log-Lik: -23268.745, Max-Change: 0.00039Iteration: 611, Log-Lik: -23268.745, Max-Change: 0.00011Iteration: 612, Log-Lik: -23268.745, Max-Change: 0.00027Iteration: 613, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 614, Log-Lik: -23268.745, Max-Change: 0.00037Iteration: 615, Log-Lik: -23268.745, Max-Change: 0.00053Iteration: 616, Log-Lik: -23268.745, Max-Change: 0.00039Iteration: 617, Log-Lik: -23268.745, Max-Change: 0.00011Iteration: 618, Log-Lik: -23268.745, Max-Change: 0.00027Iteration: 619, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 620, Log-Lik: -23268.745, Max-Change: 0.00037Iteration: 621, Log-Lik: -23268.745, Max-Change: 0.00011Iteration: 622, Log-Lik: -23268.745, Max-Change: 0.00046Iteration: 623, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 624, Log-Lik: -23268.745, Max-Change: 0.00037Iteration: 625, Log-Lik: -23268.745, Max-Change: 0.00019Iteration: 626, Log-Lik: -23268.745, Max-Change: 0.00047Iteration: 627, Log-Lik: -23268.745, Max-Change: 0.00015Iteration: 628, Log-Lik: -23268.745, Max-Change: 0.00013Iteration: 629, Log-Lik: -23268.745, Max-Change: 0.00033Iteration: 630, Log-Lik: -23268.745, Max-Change: 0.00047Iteration: 631, Log-Lik: -23268.745, Max-Change: 0.00036Iteration: 632, Log-Lik: -23268.745, Max-Change: 0.00010Iteration: 633, Log-Lik: -23268.745, Max-Change: 0.00025Iteration: 634, Log-Lik: -23268.745, Max-Change: 0.00013Iteration: 635, Log-Lik: -23268.745, Max-Change: 0.00034Iteration: 636, Log-Lik: -23268.745, Max-Change: 0.00049Iteration: 637, Log-Lik: -23268.745, Max-Change: 0.00036Iteration: 638, Log-Lik: -23268.745, Max-Change: 0.00010Iteration: 639, Log-Lik: -23268.745, Max-Change: 0.00025Iteration: 640, Log-Lik: -23268.745, Max-Change: 0.00014Iteration: 641, Log-Lik: -23268.745, Max-Change: 0.00034Iteration: 642, Log-Lik: -23268.745, Max-Change: 0.00010

#> now performing linear transformation to match anchor with SL method

out_link$constants   # transformation constants

#>         A         B 
#>  1.037453 -0.156256 

out_link$ipar_linked # item parameters linked to anchor

#>                 a           b1           b2       b3       b4
#> EDANX01 3.2958037  0.332945117  1.111039156 1.902965 2.742267
#> EDANX02 3.0622238  0.433288645  1.315080401 2.196475 3.219116
#> EDANX03 3.2979593  0.523163105  1.192102610 2.038601 2.739395
#> EDANX05 3.4324088 -0.194097681  0.609040037 1.411965 2.295667
#> EDANX07 2.9993302  0.602830830  1.037890755 1.888791 2.455920
#> EDANX08 2.4579614  0.356556237  1.058842565 1.820244 2.644322
#> EDANX12 2.5804918 -0.465525442  0.449929258 1.682292 2.952394
#> EDANX13 1.5683700  0.493436205  1.415485086 2.817910 4.091944
#> EDANX16 1.7530059  0.148510315  1.020179999 2.127068 3.236823
#> EDANX18 3.8644713  0.518667615  1.203650366 1.933225 2.742653
#> EDANX20 1.6906661  0.101051665  1.267951524 2.231318 3.165508
#> EDANX21 2.0017432 -0.313086412  0.758465868 1.933074 2.964460
#> EDANX24 1.8564287  0.296560176  0.885063380 1.682587 2.981692
#> EDANX26 2.0278106 -0.058672438  0.817285897 1.833177 3.212523
#> EDANX27 2.6464821  0.401503496  1.134509184 1.986363 2.819561
#> EDANX30 3.0164915 -0.438181656  0.391458937 1.308866 2.437926
#> EDANX33 3.2486404  0.938350378  1.713392885 2.477817 3.172492
#> EDANX37 1.8308998 -0.246345039  0.666411133 1.703958 2.551484
#> EDANX40 3.4149142  0.523184824  1.284797215 2.190037 2.921980
#> EDANX41 3.0452185  0.344377995  0.992878723 1.796809 2.570575
#> EDANX44 1.2357789  0.568954917  1.573298454 3.038949 4.605740
#> EDANX46 3.7435838 -0.168665246  0.657189545 1.638994 2.870027
#> EDANX47 2.1574875 -0.255380395  0.626990461 1.653792 2.741561
#> EDANX48 2.7493987 -0.180619678  0.692771114 1.596422 2.433739
#> EDANX49 1.3377445 -0.940554419 -0.001145037 1.250150 2.514736
#> EDANX51 2.4373836 -0.413742134  0.411223829 1.423813 2.623174
#> EDANX53 3.6498464 -0.169427056  0.609333541 1.555401 2.519724
#> EDANX54 2.9205369 -0.427030826  0.344041817 1.331289 2.453849
#> EDANX55 3.3299260  0.247712845  1.020432625 1.901245 2.788810
#> MASQ1   2.4979226  0.630902076  1.748393944 2.537353 4.333179
#> MASQ2   0.7814195  1.223432408  3.116857909 4.660312 7.916064
#> MASQ3   3.0317190 -0.032452457  1.183205123 1.857124 3.022487
#> MASQ4   3.1094048 -0.105864366  1.047902165 1.755442 2.620062
#> MASQ5   1.2432106  1.621522509  3.249574154 3.825610 4.812281
#> MASQ6   1.0932594  0.248342417  1.997733438 3.187117 5.521528
#> MASQ7   2.3249659 -0.109136222  1.012716151 1.738695 2.734497
#> MASQ8   2.1957114 -0.006593479  1.069396645 1.764179 2.773100
#> MASQ9   1.1716669  0.976809734  2.497283669 3.678265 5.242981
#> MASQ10  2.3649432  0.036914359  1.060579141 1.630818 2.622806
#> MASQ11  0.8284392 -1.117628847  1.015427520 2.771471 4.887266

out_link <- runLinking(data_asq, "FIXEDPAR")

#> response data has 40 items

#> performing fixed parameter calibration, fixing 29 items from anchor data

#> Iteration: 1, Log-Lik: -24053.720, Max-Change: 2.73844Iteration: 2, Log-Lik: -23453.207, Max-Change: 0.72611Iteration: 3, Log-Lik: -23439.758, Max-Change: 0.41693Iteration: 4, Log-Lik: -23438.315, Max-Change: 0.66336Iteration: 5, Log-Lik: -23437.791, Max-Change: 0.50563Iteration: 6, Log-Lik: -23437.683, Max-Change: 0.13141Iteration: 7, Log-Lik: -23437.680, Max-Change: 0.08747Iteration: 8, Log-Lik: -23437.677, Max-Change: 0.00486Iteration: 9, Log-Lik: -23437.677, Max-Change: 0.00021Iteration: 10, Log-Lik: -23437.677, Max-Change: 0.00015Iteration: 11, Log-Lik: -23437.677, Max-Change: 0.00013Iteration: 12, Log-Lik: -23437.677, Max-Change: 0.00002

out_link$ipar_linked # item parameters linked to anchor

#>                 a          b1         b2       b3       b4
#> EDANX01 3.6021544  0.34159048 1.08952607 1.960101 2.698704
#> EDANX02 3.4504105  0.49115372 1.33003913 2.155174 2.880245
#> EDANX03 3.3365573  0.62131513 1.24324827 2.061498 2.758621
#> EDANX05 3.3552816 -0.18966149 0.59808268 1.574890 2.445823
#> EDANX07 3.5509265  0.53941950 1.04602617 1.865002 2.384703
#> EDANX08 2.8482390  0.36804661 1.06412460 1.853262 2.544713
#> EDANX12 2.7344963 -0.56874342 0.47212997 1.671720 2.969534
#> EDANX13 1.4299890  0.32031811 1.38525972 2.740602 4.657334
#> EDANX16 1.8026250 -0.03826752 0.87866300 2.112580 3.045698
#> EDANX18 2.9857713  0.56615423 1.30206069 2.149592 3.060772
#> EDANX20 1.7235167  0.01038630 1.16848655 2.202962 3.108540
#> EDANX21 2.2320871 -0.30039122 0.73730822 1.859789 2.924760
#> EDANX24 1.8300229  0.33186966 0.98209963 1.887270 3.023501
#> EDANX26 1.9506502 -0.02653687 0.87921023 2.014506 3.240672
#> EDANX27 2.8602293  0.43614207 1.12685718 2.029540 2.777844
#> EDANX30 3.0318433 -0.51540331 0.31555223 1.351650 2.299683
#> EDANX33 2.5819082  1.14704504 1.81665694 2.716464 3.589089
#> EDANX37 1.7091745 -0.03958087 0.88362788 2.001631 3.104371
#> EDANX40 3.8831993  0.48588040 1.26317743 2.111117 2.898513
#> EDANX41 3.6595132  0.36439509 1.03380800 1.780490 2.621156
#> EDANX44 1.2704961  0.57192503 1.45897673 2.677899 3.972340
#> EDANX46 3.3981373 -0.21663501 0.63213040 1.644448 2.731239
#> EDANX47 2.5365096 -0.23896204 0.67780928 1.795603 2.900682
#> EDANX48 3.0433412 -0.33156350 0.55586628 1.455912 2.339870
#> EDANX49 1.5157005 -0.82984231 0.08690911 1.167730 2.379728
#> EDANX51 2.4116611 -0.46182440 0.39835327 1.363589 2.402821
#> EDANX53 3.6561060 -0.23177839 0.59522713 1.563458 2.499054
#> EDANX54 3.3503266 -0.50940366 0.31066694 1.250223 2.296618
#> EDANX55 3.1273340  0.07436260 0.94088815 1.847720 2.775704
#> MASQ1   2.4553497  0.62790104 1.76305752 2.556936 4.346699
#> MASQ2   0.7737661  1.22523313 3.13716697 4.695677 7.982061
#> MASQ3   2.9565331 -0.04681064 1.19269227 1.877235 3.041972
#> MASQ4   3.0397031 -0.12152483 1.05359287 1.773160 2.642030
#> MASQ5   1.2272508  1.63114624 3.27898792 3.861737 4.859650
#> MASQ6   1.0794218  0.24091539 2.01114033 3.214757 5.576666
#> MASQ7   2.2877344 -0.12364517 1.01625401 1.752663 2.756741
#> MASQ8   2.1625165 -0.01950077 1.07289006 1.777268 2.794920
#> MASQ9   1.1514936  0.98152511 2.52516382 3.723843 5.312661
#> MASQ10  2.3261540  0.02420664 1.06489768 1.643662 2.644612
#> MASQ11  0.8225675 -1.13444246 1.01296274 2.781956 4.912498

# }