Compute Crosswalk Tables

runRSSS is a function for generating raw-score to standard-score crosswalk tables from supplied calibrated item parameters.

runRSSS(
  data,
  ipar_linked,
  prior_mean = 0,
  prior_sd = 1,
  min_theta = -4,
  max_theta = 4,
  inc = 0.05,
  min_score = 1
)

Arguments

data

a PROsetta_data object. See loadData for loading a dataset.

ipar_linked

an object returned from runLinking or runCalibration.

prior_mean

prior mean. (default = 0.0)

prior_sd

prior standard deviation. (default = 1.0)

min_theta

the lower limit of theta quadratures for numerical integration. (default = -4)

max_theta

the upper limit of theta quadratures for numerical integration. (default = 4)

inc

the increment between theta quadratures for numerical integration. (default = 0.05)

min_score

minimum item score (0 or 1) for each scale (1, 2, and combined). If a single value is supplied, the value is applied to all scales. (default = 1)

Value

runRSSS returns a list containing crosswalk tables.

Examples

# \donttest{
out_link    <- runLinking(data_asq, method = "FIXEDPAR")
score_table <- runRSSS(data_asq, out_link)
# }