ASQ dataset

This dataset is associated with the following objects:

Details

response_asq a data.frame containing raw response data of 751 participants and 41 variables.

prosettaid participant IDs.

EDANX01 -- MASQ11 response to items.

itemmap_asq a data.frame containing the item map, describing the items in each instrument.

item_order item numeric IDs. This column refers to the column item_order in anchor item attributes.

instrument the instrument ID that each item belongs to.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

item_name new item ID strings to be used in the combined scale.

ncat the number of response categories.

anchor_asq a data.frame containing anchor item parameters for 29 items.

item_order item numeric IDs.

item_id item ID strings. This column refers to column names in raw response data, excluding the participant ID column.

a the discrimination (slope) parameter for the graded response model.

cb1 - cb4 the boundaries between each category-pair for the graded response model.

data_asq a PROsetta_data object containing the datasets above. See loadData for creating PROsetta_data objects.

Examples

## load datasets into a PROsetta_data object
data_asq <- loadData(
  response = response_asq,
  itemmap  = itemmap_asq,
  anchor   = anchor_asq
)

## run descriptive statistics
runDescriptive(data_asq)

## run item parameter calibration on the response data, linking to the anchor item parameters
runLinking(data_asq, method = "FIXEDPAR")