Run Test Equating

runEquateObserved is a function to perform equipercentile test equating between two scales. A concordance table is produced, mapping the observed raw scores from one scale to the scores from another scale.

runEquateObserved(
  data,
  scale_from = 2,
  scale_to = 1,
  type_to = "raw",
  rsss = NULL,
  eq_type = "equipercentile",
  smooth = "loglinear",
  degrees = list(3, 1),
  boot = TRUE,
  reps = 100,
  ...
)

Arguments

data
a PROsetta_data object. See loadData for loading a dataset.

scale_from
the scale ID of the input scale. References to itemmap in data argument.

scale_to
the scale ID of the target scale to equate to. References to itemmap in data argument.

type_to
the type of score to use in the target scale frequency table. Accepts 'raw' (default) and 'tscore'. 'tscore' requires RSSS table to be supplied.

rsss
the RSSS table to use to map each raw score level onto t-score. See runRSSS.

eq_type
the type of equating to be passed onto equate in 'equate' package.

smooth
the type of smoothing method to be passed onto presmoothing in 'equate' package.

degrees
the degrees of smoothing to be passed onto presmoothing.

boot
performs bootstrapping if TRUE.

reps
the number of replications to perform in bootsrapping.

...
other arguments to pass onto equate.

Value

runEquateObserved returns an equate object containing the test equating result.

The printed summary statistics indicate the distributional properties of the two supplied scales and the equated scale.

x corresponds to scale_from.

y corresponds to scale_to.

yx corresponds to scale_from after equating to scale_to.

See equate for details.

The concordance table is stored in concordance slot.

Examples

out_eq_raw <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  eq_type = "equipercentile", smooth = "loglinear"
)

#> runEquateObserved requires complete data, attempting to remove cases

#> getCompleteData: filtered 8 cases with missing responses in all scales

#> performing loglinear presmoothing on scale 2 (scale_from) distribution

#> performing loglinear presmoothing on scale 1 (scale_to) distribution

out_eq_raw$concordance

#>    raw_2     raw_1   raw_1_se raw_1_se_boot
#> 1     11  29.71858  0.2483603     0.1776580
#> 2     12  32.33361  0.4663568     0.4567534
#> 3     13  35.10463  0.6681376     0.6431408
#> 4     14  38.02455  0.8662512     0.8745018
#> 5     15  41.09143  1.0609494     1.1151137
#> 6     16  44.29367  1.2524657     1.2376782
#> 7     17  47.61721  1.5077675     1.4708558
#> 8     18  51.04427  1.6805879     1.7977498
#> 9     19  54.53510  1.9308974     2.0485023
#> 10    20  58.07716  2.0793550     2.2498398
#> 11    21  61.62444  2.3104050     2.4505889
#> 12    22  65.16029  2.4379903     2.7390007
#> 13    23  68.65089  2.6546378     2.8919124
#> 14    24  72.08339  2.7663267     3.0725263
#> 15    25  75.43304  2.8846186     3.1458178
#> 16    26  78.69900  3.1006498     3.1075550
#> 17    27  81.86936  3.2176468     3.2448089
#> 18    28  84.94022  3.3460076     3.2878864
#> 19    29  87.91190  3.4893302     3.2798016
#> 20    30  90.78567  3.6547336     3.5960570
#> 21    31  93.56298  3.8541951     4.1598912
#> 22    32  96.25648  3.8828176     4.2439661
#> 23    33  98.86327  4.1245578     4.2912381
#> 24    34 101.38572  4.2016116     4.4460155
#> 25    35 103.84034  4.5176063     4.5052509
#> 26    36 106.22205  4.6330766     4.7796254
#> 27    37 108.53555  5.0946338     5.1840713
#> 28    38 110.80177  5.2584873     5.7175696
#> 29    39 113.01125  5.4650506     6.5375165
#> 30    40 115.17125  5.7131451     6.7778803
#> 31    41 117.28871  6.0029191     7.1314362
#> 32    42 119.37049  6.3357393     7.5557551
#> 33    43 121.42353  6.7139238     8.2436651
#> 34    44 123.45521  7.1402725     9.4558585
#> 35    45 125.47375  7.6172335     9.9570781
#> 36    46 127.48883  8.1454867    10.0979919
#> 37    47 129.51373  9.6756492    10.9796842
#> 38    48 131.56670 10.3680795    10.9644244
#> 39    49 133.67002 11.0295096     9.7580908
#> 40    50 135.85632 11.5574281     9.5228130
#> 41    51 138.17184 11.7867191     8.5194702
#> 42    52 140.70935 12.8142324     7.0466973
#> 43    53 143.65612 10.7336688     4.4570932

out_link <- runLinking(data_asq, method = "FIXEDPAR")

#> response data has 40 items

#> performing fixed parameter calibration, fixing 29 items from anchor data

#> Iteration: 1, Log-Lik: -24053.720, Max-Change: 2.73844Iteration: 2, Log-Lik: -23453.207, Max-Change: 0.72611Iteration: 3, Log-Lik: -23439.758, Max-Change: 0.41693Iteration: 4, Log-Lik: -23438.315, Max-Change: 0.66336Iteration: 5, Log-Lik: -23437.791, Max-Change: 0.50563Iteration: 6, Log-Lik: -23437.683, Max-Change: 0.13141Iteration: 7, Log-Lik: -23437.680, Max-Change: 0.08747Iteration: 8, Log-Lik: -23437.677, Max-Change: 0.00486Iteration: 9, Log-Lik: -23437.677, Max-Change: 0.00021Iteration: 10, Log-Lik: -23437.677, Max-Change: 0.00015Iteration: 11, Log-Lik: -23437.677, Max-Change: 0.00013Iteration: 12, Log-Lik: -23437.677, Max-Change: 0.00002

out_rsss <- runRSSS(data_asq, out_link)
out_eq_tscore <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  type_to = "tscore", rsss = out_rsss,
  eq_type = "equipercentile", smooth = "loglinear"
)

#> runEquateObserved requires complete data, attempting to remove cases

#> getCompleteData: filtered 8 cases with missing responses in all scales

#> performing loglinear presmoothing on scale 2 (scale_from) distribution

#> mapping scale 1 (scale_to) raw scores to t-scores using supplied rsss

#> performing loglinear presmoothing on scale 1 (scale_to) distribution

out_eq_tscore$concordance

#>    raw_2 tscore_1 tscore_1_se tscore_1_se_boot
#> 1     11 33.36209   0.1839703        0.6446305
#> 2     12 39.16707   0.4815614        0.6062406
#> 3     13 42.67596   0.7319421        0.7372714
#> 4     14 45.26573   0.9458103        0.6846322
#> 5     15 47.32729   1.1383102        0.6787958
#> 6     16 49.14145   1.3681190        0.7068874
#> 7     17 50.67372   1.5143645        0.7161785
#> 8     18 52.06602   1.7166059        0.7645926
#> 9     19 53.42760   1.8598942        0.7430162
#> 10    20 54.63619   2.0589723        0.7247637
#> 11    21 55.80801   2.1734974        0.7232591
#> 12    22 56.91759   2.3009457        0.7320533
#> 13    23 58.01370   2.5120668        0.7590891
#> 14    24 59.09281   2.6510670        0.8509649
#> 15    25 60.13891   2.8035007        0.8998454
#> 16    26 61.10101   2.9318021        1.0102552
#> 17    27 62.13006   3.1076892        1.1571391
#> 18    28 63.08775   3.2529137        1.2346685
#> 19    29 64.03669   3.4537848        1.2695190
#> 20    30 65.01772   3.6716763        1.3646616
#> 21    31 65.91675   3.8526120        1.3634824
#> 22    32 66.82541   4.1112424        1.3803004
#> 23    33 67.75277   4.4035371        1.5358781
#> 24    34 68.66372   4.7417109        1.6947595
#> 25    35 69.49223   4.7656345        1.8148635
#> 26    36 70.34646   5.1343513        1.8397188
#> 27    37 71.18831   5.2697067        1.8888000
#> 28    38 72.02319   5.6952262        1.9958218
#> 29    39 72.81581   5.8412341        2.1047648
#> 30    40 73.62047   6.0185681        2.0809325
#> 31    41 74.40101   6.5274404        2.6500516
#> 32    42 75.20521   6.8170078        2.6910596
#> 33    43 75.95228   6.9870905        2.6784137
#> 34    44 76.75614   7.2980791        3.1055265
#> 35    45 77.53426   7.6129267        4.1506073
#> 36    46 78.34520   8.0978153        4.4562266
#> 37    47 79.17099   8.6047147        5.1815332
#> 38    48 80.07766   8.2529774        5.3458621
#> 39    49 81.06941   8.9067580        5.2383716
#> 40    50 82.19355   9.6964121        4.9037425
#> 41    51 83.54347  10.4752764        4.2583387
#> 42    52 85.28985  11.1363085        3.6029261
#> 43    53 87.42110  10.8919224        2.3520967