Run Test Equating

runEquateObserved is a function to perform equipercentile test equating between two scales. A concordance table is produced, mapping the observed raw scores from one scale to the scores from another scale.

runEquateObserved(
  data,
  scale_from = 2,
  scale_to = 1,
  type_to = "raw",
  rsss = NULL,
  eq_type = "equipercentile",
  smooth = "loglinear",
  degrees = list(3, 1),
  boot = TRUE,
  reps = 100,
  verbose = FALSE,
  ...
)

Arguments

data

a PROsetta_data object. See loadData for loading a dataset.

scale_from

the scale ID of the input scale. References to itemmap in data argument. (default = 2)

scale_to

the scale ID of the target scale to equate to. References to itemmap in data argument. (default = 1)

type_to

the type of score to use in the target scale frequency table. Accepts raw, tscore, and theta. tscore and theta require argument rsss to be supplied. (default = raw)

rsss

the RSSS table to use to map each raw score level onto a t-score or a theta. See runRSSS.

eq_type

the type of equating to be passed onto equate in 'equate' package. (default = equipercentile)

smooth

the type of smoothing method to be passed onto presmoothing in 'equate' package. (default = loglinear)

degrees

the degrees of smoothing to be passed onto presmoothing. (default = list(3, 1))

boot

performs bootstrapping if TRUE. (default = TRUE)

reps

the number of replications to perform in bootstrapping. (default = 100)

verbose

if TRUE, print status messages. (default = FALSE)

...

other arguments to pass onto equate.

Value

runEquateObserved returns an equate object containing the test equating result. The printed summary statistics indicate the distributional properties of the two supplied scales and the equated scale.

x corresponds to scale_from.

y corresponds to scale_to.

yx corresponds to scale_from after equating to scale_to.

See equate for details. The concordance table is stored in concordance slot.

Examples

out_eq_raw <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  eq_type = "equipercentile", smooth = "loglinear"
)
out_eq_raw$concordance

# \donttest{
out_link <- runLinking(data_asq, method = "FIXEDPAR")
out_rsss <- runRSSS(data_asq, out_link)
out_eq_tscore <- runEquateObserved(data_asq,
  scale_to = 1, scale_from = 2,
  type_to = "tscore", rsss = out_rsss,
  eq_type = "equipercentile", smooth = "loglinear"
)
out_eq_tscore$concordance
# }