Run Scale Linking

runLinking is a function for obtaining item parameters from the response data in the metric of supplied anchor item parameters.

runLinking(data, method, verbose = FALSE, ...)

Arguments

data

a PROsetta_data object. See loadData for loading a dataset.

method

the linking method to use. Accepts:

MM for mean-mean method

MS for mean-sigma method

HB for Haebara method

SL for Stocking-Lord method

FIXEDPAR for fixed parameter calibration

CP for calibrated projection using fixed parameter calibration on the anchor dimension

CPLA for linear approximation of calibrated projection. This is identical to 'CP' but uses approximation in runRSSS

CPFIXEDDIM for calibrated projection using mean and variance constraints on the anchor dimension

Linear transformation methods (i.e., MM, MS, HB, SL) are performed with plink in 'plink' package.

verbose

if TRUE, print status messages. (default = FALSE)

...

additional arguments to pass onto mirt in 'mirt' package.

Value

runLinking returns a list containing the scale linking results.

constants linear transformation constants. Only available when linear transformation methods were used (i.e., MM, MS, HB, SL).

ipar_linked item parameters calibrated to the response data, and linked to the metric of anchor item parameters.

ipar_anchor anchor item parameters used in linking.

Examples

# \donttest{
out_link <- runLinking(data_asq, "SL", technical = list(NCYCLES = 1000))
out_link$constants   # transformation constants
out_link$ipar_linked # item parameters linked to anchor
out_link <- runLinking(data_asq, "FIXEDPAR")
out_link$ipar_linked # item parameters linked to anchor
# }